Modularize CLI and implement PINN-ready ground truth generation
- Split CLI/aircraft_diffusion_cfd.py into modular components (<500 lines each)
- Add GroundTruthExporter for 3D field-level data saving (velocity/pressure)
- Refactor D3Q27 solver for force source transparency (Physical vs Proxy)
- Implement robust OpenFOAM force.dat parsing for independent validation
- Add accuracy-benchmark CLI command for surrogate validation
- Update tests and GUI integration for new modular structure
- Remove legacy cleanup files (test_fix.py, todo.md)
