--- 20mm_cube.stl
@@ -1,86 +1,0 @@
-solid model
-facet normal 0.0 0.0 -1.0
-outer loop
-vertex 20.0 0.0 0.0
-vertex 0.0 -20.0 0.0
-vertex 0.0 0.0 0.0
-endloop
-endfacet
-facet normal 0.0 0.0 -1.0
-outer loop
-vertex 0.0 -20.0 0.0
-vertex 20.0 0.0 0.0
-vertex 20.0 -20.0 0.0
-endloop
-endfacet
-facet normal -0.0 -1.0 -0.0
-outer loop
-vertex 20.0 -20.0 20.0
-vertex 0.0 -20.0 0.0
-vertex 20.0 -20.0 0.0
-endloop
-endfacet
-facet normal -0.0 -1.0 -0.0
-outer loop
-vertex 0.0 -20.0 0.0
-vertex 20.0 -20.0 20.0
-vertex 0.0 -20.0 20.0
-endloop
-endfacet
-facet normal 1.0 0.0 0.0
-outer loop
-vertex 20.0 0.0 0.0
-vertex 20.0 -20.0 20.0
-vertex 20.0 -20.0 0.0
-endloop
-endfacet
-facet normal 1.0 0.0 0.0
-outer loop
-vertex 20.0 -20.0 20.0
-vertex 20.0 0.0 0.0
-vertex 20.0 0.0 20.0
-endloop
-endfacet
-facet normal -0.0 -0.0 1.0
-outer loop
-vertex 20.0 -20.0 20.0
-vertex 0.0 0.0 20.0
-vertex 0.0 -20.0 20.0
-endloop
-endfacet
-facet normal -0.0 -0.0 1.0
-outer loop
-vertex 0.0 0.0 20.0
-vertex 20.0 -20.0 20.0
-vertex 20.0 0.0 20.0
-endloop
-endfacet
-facet normal -1.0 0.0 0.0
-outer loop
-vertex 0.0 0.0 20.0
-vertex 0.0 -20.0 0.0
-vertex 0.0 -20.0 20.0
-endloop
-endfacet
-facet normal -1.0 0.0 0.0
-outer loop
-vertex 0.0 -20.0 0.0
-vertex 0.0 0.0 20.0
-vertex 0.0 0.0 0.0
-endloop
-endfacet
-facet normal -0.0 1.0 0.0
-outer loop
-vertex 0.0 0.0 20.0
-vertex 20.0 0.0 0.0
-vertex 0.0 0.0 0.0
-endloop
-endfacet
-facet normal -0.0 1.0 0.0
-outer loop
-vertex 20.0 0.0 0.0
-vertex 0.0 0.0 20.0
-vertex 20.0 0.0 20.0
-endloop
-endfacet
-endsolid model
