+++ CLI/aircraft_diffusion_cfd.py
@@ -1,2280 +1,156 @@
 #!/usr/bin/env python3
 """
 Aircraft Structural Design via Diffusion Models + FluidX3D CFD
-Combines TRM/HRM principles with diffusion-based 3D voxel generation,
-GPU-accelerated CFD simulation, and marching cubes STL export.
-
-Fully optimized for 8-13GB VRAM with pipelined training and inference.
-TRM/HRM Recursive Style Implementation with:
-- FluidX3D integration with adaptive mesh refinement
-- 4-step consistency model distillation
-- Grouped-query attention (4 groups, 50% KV-cache reduction)
-- Gradient checkpointing (60% VRAM savings)
-- Pipeline parallelism for CFD/diffusion overlap
+(Split module version)
 """
 
 import os
 import sys
-import json
-import pickle
-import argparse
-import warnings
-import subprocess
-import tempfile
-import threading
-import multiprocessing as mp
-from pathlib import Path
-from typing import Dict, List, Tuple, Optional, Any, Union
-from dataclasses import dataclass, asdict
-from datetime import datetime
-from concurrent.futures import ThreadPoolExecutor, ProcessPoolExecutor
-import asyncio
-
-import numpy as np
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from torch.utils.data import Dataset, DataLoader
-from torch.optim import AdamW
-from torch.optim.lr_scheduler import CosineAnnealingLR
-from torch.utils.tensorboard import SummaryWriter
-from torch.cuda.amp import GradScaler, autocast
-
-from tqdm import tqdm
-import yaml
-from scipy.ndimage import label, binary_dilation
-from skimage import measure
-import trimesh
-from advanced_lbm_solver import D3Q27CascadedSolver
-
-warnings.filterwarnings('ignore')
-
-OPENFOAM_ROOT = Path(os.environ.get("OPENFOAM_ROOT", "/home/darsh/.openclaw/openfoam/usr/share/openfoam"))
-OPENFOAM_BIN = OPENFOAM_ROOT / "bin"
-OPENFOAM_AVAILABLE = all((OPENFOAM_BIN / cmd).exists() for cmd in ("blockMesh", "snappyHexMesh", "simpleFoam"))
-
-# ============================================================================
-# CONFIG & DATACLASSES
-# ============================================================================
-
-@dataclass
-class DiffusionConfig:
-    """Diffusion model hyperparameters with consistency distillation support"""
-    timesteps: int = 100
-    beta_start: float = 0.0001
-    beta_end: float = 0.02
-    sampling_timesteps: int = 250
-    guidance_scale: float = 7.5
-    # Consistency distillation settings
-    teacher_steps: int = 2000  # Original teacher model steps
-    student_steps: int = 32     # Target student model steps
-    progressive_distillation: List[int] = None  # 500→250→125→64→32→16→8→4
-    
-    def __post_init__(self):
-        if self.progressive_distillation is None:
-            self.progressive_distillation = [500, 250, 125, 64, 32]
-    
-@dataclass
-class ModelConfig:
-    """Model architecture parameters with grouped-query attention"""
-    latent_dim: int = 16
-    xyz_dim: int = 3
-    encoder_channels: List[int] = None
-    decoder_channels: List[int] = None
-    # Grouped-query attention instead of multi-head
-    attention_groups: int = 8  # 4 groups instead of 8 heads (50% KV-cache reduction)
-    attention_kv_groups: int = 8  # Groups for key/value
-    num_attention_layers: int = 4
-    # Grid resolution - configurable for different lattice sizes
-    base_grid_resolution: int = 32  # Consistent grid resolution for voxel, CFD, etc.
-    grid_resolution: int = None  # Working grid resolution (defaults to base_grid_resolution if not set)
-    target_grid_resolution: int = 1024 # Final high-res target
-    # Memory optimization
-    enable_gradient_checkpointing: bool = True  # 60% VRAM savings
-    use_torch_compile: bool = False  # Kernel fusion
-
-    def __post_init__(self):
-        if self.encoder_channels is None:
-            self.encoder_channels = [24, 32, 48]
-        if self.decoder_channels is None:
-            self.decoder_channels = [48, 32, 24]
-        # Set working grid resolution if not specified
-        if self.grid_resolution is None:
-            self.grid_resolution = self.base_grid_resolution
-
-@dataclass
-class TrainingConfig:
-    """Training hyperparameters"""
-    batch_size: int = 4
-    learning_rate: float = 2e-4
-    weight_decay: float = 1e-4
-    num_epochs: int = 100
-    warmup_steps: int = 1000
-    gradient_clip: float = 0.99
-    ema_decay: float = 0.99
-    disconnection_penalty: float = 50.0
-    precision: str = 'float32'
-    save_interval: int = 5
-    val_interval: int = 2
-    # Pipeline parallelism
-    enable_pipeline_parallelism: bool = True  # Overlap CFD with diffusion
-    num_pipeline_stages: int = 8  # CFD + Diffusion stages
-
-@dataclass
-class LBMPhysicsConfig:
-    """Easy-to-update configuration for LBM physics constants"""
-    # Turbulence modeling
-    turbulence_model: str = "dynamic_smagorinsky"  # Options: 'smagorinsky', 'dynamic_smagorinsky', 'wale', 'none'
-    smagorinsky_constant: float = 0.17  # Cs for LES turbulence model
-    wale_constant: float = 0.5  # Cw for WALE model
-    use_les_turbulence: bool = True
-    physical_length_scale: float = 1.0  # Physical length of the voxel grid (m)
-    grid_spacing: float = 0.01  # Physical spacing per grid cell (m) - calculated from physical_length_scale
-    time_step: float = 0.001  # Time step size (s)
-    test_filter_ratio: float = 2.0  # Ratio for test filter in dynamic Smagorinsky model
-    dynamic_cs_clip_min: float = 0.0  # Minimum Cs value for dynamic model
-    dynamic_cs_clip_max: float = 0.2  # Maximum Cs value for dynamic model
-    use_vorticity_confinement: bool = True  # Enable vorticity confinement
-    vc_adaptive_strength: float = 0.1  # Adaptive vorticity confinement strength
-    vc_adaptive: bool = True  # Adaptive strength factor
-    vorticity_confinement_epsilon: float = 0.1  # Vorticity confinement parameter
-    compute_q_criterion: bool = True  # Compute Q-criterion for vortex identification
-    q_threshold: float = 0.0  # Threshold for Q-criterion visualization
-
-    # MRT relaxation times (for different moment components)
-    # s0-s18 correspond to different moments in MRT collision
-    # Lower values = more dissipation, higher stability
-    s_nu: float = None  # Set from Reynolds number (kinematic viscosity)
-    s_bulk: float = 1.0  # Bulk viscosity relaxation (density fluctuations)
-    s_energy: float = 1.2  # Energy mode relaxation
-    s_higher: float = 1.4  # Higher order moments relaxation
-
-    # D3Q27 Cascaded LBM relaxation parameters
-    s_nu_d3q27: float = 1.0 / 0.6    # Viscosity relaxation
-    s_e_d3q27: float = 1.2           # Energy relaxation
-    s_h_d3q27: float = 1.6           # Higher order relaxation
-    tau_min_d3q27: float = 0.52      # BGK stability floor for under-resolved high-Re runs
-
-    # Boundary conditions
-    inlet_velocity_relaxation: float = 0.5  # For Zou-He BC smoothing
-    convergence_tolerance: float = 1e-5  # Velocity change threshold
-    max_iterations: int = 50000  # Safety limit
-    check_convergence_every: int = 250  # Steps between convergence checks
-
-    # Stability
-    max_mach: float = 0.3  # Maximum Mach number for stability
-    target_lattice_velocity: float = 0.12  # Cap inlet speed in lattice units for high-resolution stability
-    cfl_safety_factor: float = 0.8  # CFL condition safety margin
-
-    # Low-Mach corrections
-    use_incompressible_correction: bool = True  # Use rho0=1 for incompressible
-
-    # Force computation
-    momentum_exchange_correction: bool = True  # Apply momentum-exchange method
-    use_triton_streaming: bool = False  # Keep disabled until fused kernel matches physics path exactly
-    drag_link_metric_exponent: Optional[float] = None  # Auto D3Q27 face/edge/corner metric correction
-    drag_reference_speed: float = 80.0  # Natural-unit reference speed for projected-pressure Cd labels
-    drag_speed_normalization_exponent: float = 1.0  # OpenFOAM pressure fallback scales nearly linearly with U_inf
-    use_shape_drag_correction: bool = True
-    shape_drag_correction_coefficients: Tuple[float, ...] = (
-        -12.633030612111941, 27.87582461044955, -10.247055184812014,
-        22.962648171191816, -17.337224317584685, -3.946645931513679,
-        0.08323209768046214, 4.548014973469924, -5.179313884992105,
-        -7.623947231425998,
-    )
-    shape_drag_correction_min: float = 0.1
-    shape_drag_correction_max: float = 3.0
-    
-@dataclass
-class CFDConfig:
-    """FluidX3D simulation parameters with adaptive mesh refinement"""
-    solver_type: str = "D3Q27"
-    base_grid_resolution: int = 32  # Consistent grid resolution - no resizing needed
-    mach_number: float = 0.025
-    reynolds_number: float = 1e6
-    simulation_steps: int = 1000
-    output_interval: int = 50
-    device_id: int = 0
-    # Adaptive mesh refinement
-    use_amr: bool = False
-    adaptive_cells_target: int = int(5e3)  # Target ~5k cells for AMR
-    refinement_levels: int = 3
-    # LBM configuration
-    lbm_config: LBMPhysicsConfig = None   # LBM parameters
-    # Backwards compatibility parameter - default to base_grid_resolution
-    resolution: int = None  # If provided, sets base_grid_resolution
-
-    def __post_init__(self):
-        # Set default resolution if not provided
-        if self.resolution is None:
-            self.resolution = self.base_grid_resolution
-
-        # Handle backwards compatibility for resolution parameter
-        if self.resolution is not None:
-            self.base_grid_resolution = self.resolution
-
-        if self.lbm_config is None:
-            # Calculate physical grid spacing: physical_length / resolution
-            self.lbm_config = LBMPhysicsConfig()
-            self.lbm_config.grid_spacing = self.lbm_config.physical_length_scale / self.base_grid_resolution
-
-@dataclass
-class DesignSpec:
-    """Aircraft design specification"""
-    target_speed: float = 7.0  # m/s
-    space_weight: float = 0.33*100
-    drag_weight: float = 0.33*100
-    lift_weight: float = 0.34*100
-    bounding_box: Tuple[int, int, int] = (64, 64, 64)
-    vital_components: np.ndarray = None
-
-# ============================================================================
-# GROUPED-QUERY ATTENTION (50% KV-CACHE REDUCTION)
-# ============================================================================
-
-class GroupedQueryAttention(nn.Module):
-    """Memory-efficient grouped-query attention for 50% KV-cache reduction"""
-    
-    def __init__(self, channels: int, num_groups: int = 4, num_kv_groups: int = 4):
-        super().__init__()
-        self.num_groups = num_groups
-        self.num_kv_groups = num_kv_groups
-        self.channels = channels
-        self.group_size = channels // num_groups
-        self.kv_group_size = channels // num_kv_groups
-        
-        self.scale = (self.group_size) ** -0.5
-        
-        # Q projections: one per group
-        self.to_q = nn.Conv3d(channels, channels, 1)
-        
-        # KV projections: shared across KV groups
-        self.to_k = nn.Conv3d(channels, self.num_kv_groups * self.kv_group_size, 1)
-        self.to_v = nn.Conv3d(channels, self.num_kv_groups * self.kv_group_size, 1)
-        
-        # Output projection
-        self.to_out = nn.Conv3d(channels, channels, 1)
-    
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        b, c, d, h, w = x.shape
-        
-        # Compute Q, K, V
-        q = self.to_q(x)  # [B, C, D, H, W]
-        k = self.to_k(x)  # [B, num_kv_groups * kv_group_size, D, H, W]
-        v = self.to_v(x)  # [B, num_kv_groups * kv_group_size, D, H, W]
-        
-        # Reshape for grouped attention
-        q = q.view(b, self.num_groups, self.group_size, d, h, w)
-        k = k.view(b, self.num_kv_groups, self.kv_group_size, d, h, w)
-        v = v.view(b, self.num_kv_groups, self.kv_group_size, d, h, w)
-        
-        # Flatten spatial dimensions for attention computation
-        q = q.view(b, self.num_groups, self.group_size, -1).transpose(-2, -1)  # [B, num_groups, N, group_size]
-        k = k.view(b, self.num_kv_groups, self.kv_group_size, -1).transpose(-2, -1)  # [B, num_kv_groups, N, kv_group_size]
-        v = v.view(b, self.num_kv_groups, self.kv_group_size, -1).transpose(-2, -1)  # [B, num_kv_groups, N, kv_group_size]
-        
-        # Expand K and V to match Q groups
-        k_expanded = k.repeat_interleave(self.num_groups // self.num_kv_groups, dim=1)
-        v_expanded = v.repeat_interleave(self.num_groups // self.num_kv_groups, dim=1)
-        
-        # Compute attention
-        sim = torch.einsum('bgqd,bgkd->bgqk', q, k_expanded) * self.scale
-        attn = sim.softmax(dim=-1)
-        
-        out = torch.einsum('bgqk,bgkd->bgqd', attn, v_expanded)
-        out = out.transpose(-2, -1).contiguous().view(b, c, d, h, w)
-        out = self.to_out(out)
-        
-        return x + out
-
-# ============================================================================
-# GRADIENT CHECKPOINTING WRAPPER (60% VRAM SAVINGS)
-# ============================================================================
-
-class GradientCheckpointingWrapper(nn.Module):
-    """Wrapper to enable gradient checkpointing for 60% VRAM savings"""
-    
-    def __init__(self, module: nn.Module, checkpoint_every: int = 1):
-        super().__init__()
-        self.module = module
-        self.checkpoint_every = checkpoint_every
-        self.call_count = 0
-    
-    def forward(self, *args, **kwargs):
-        if self.checkpoint_every > 1:
-            self.call_count += 1
-            if self.call_count % self.checkpoint_every == 0:
-                # Use gradient checkpointing
-                return torch.utils.checkpoint.checkpoint(self.module, *args, **kwargs)
-        
-        return self.module(*args, **kwargs)
-
-
-# ============================================================================
-# GPU-RESIDENT LBM SOLVER WITH SOA LAYOUT
-# ============================================================================
-# (See advanced_lbm_solver.py for full implementation)
-# ============================================================================
-# 4-STEP CONSISTENCY MODEL
-# ============================================================================
-
-class ConsistencyModel(nn.Module):
-    """4-step consistency model replacing 1000-step diffusion"""
-    
-    def __init__(self, config: ModelConfig, diffusion_config: DiffusionConfig, dtype: torch.dtype = torch.float32):
-        super().__init__()
-        self.config = config
-        self.diffusion_config = diffusion_config
-        self.student_steps = diffusion_config.student_steps  # 4 steps
-        self.teacher_steps = diffusion_config.teacher_steps  # 1000 steps
-
-        # Teacher model (large, slow) - disable torch.compile for stability
-        teacher_config = ModelConfig(
-            latent_dim=config.latent_dim,
-            encoder_channels=config.encoder_channels,
-            decoder_channels=config.decoder_channels,
-            attention_groups=config.attention_groups,
-            enable_gradient_checkpointing=config.enable_gradient_checkpointing,
-            use_torch_compile=False  # Disable torch.compile for teacher to avoid overflow errors
-        )
-        self.teacher_model = LatentDiffusionUNet(teacher_config, diffusion_config).to(dtype)
-
-        # Student model (small, fast)
-        student_config = ModelConfig(
-            latent_dim=config.latent_dim,
-            encoder_channels=[c // 2 for c in config.encoder_channels],  # Smaller
-            decoder_channels=[c // 2 for c in config.decoder_channels],
-            attention_groups=4,
-            enable_gradient_checkpointing=True,
-            use_torch_compile=False  # Disable torch.compile for student to avoid overflow errors
-        )
-        self.student_model = LatentDiffusionUNet(student_config, diffusion_config).to(dtype)
-
-        # Initialize student with teacher weights
-        self._initialize_student()
-    
-    def _initialize_student(self):
-        """Initialize student model - cannot copy from teacher due to different sizes"""
-        # Student model has smaller channels than teacher, so we initialize randomly
-        # The student will learn to match teacher outputs through consistency training
-        for param in self.student_model.parameters():
-            if param.dim() > 1:
-                nn.init.xavier_uniform_(param)
-            else:
-                nn.init.zeros_(param)
-    
-    def consistency_loss(self, x_0: torch.Tensor, t_student: torch.Tensor, t_teacher: torch.Tensor) -> torch.Tensor:
-        """Consistency training loss between teacher and student models"""
-        batch_size = x_0.shape[0]
-        device = x_0.device
-        
-        # Teacher prediction at high resolution
-        noise = torch.randn_like(x_0)
-        x_t_teacher = self._add_noise(x_0, t_teacher, noise)
-        with torch.no_grad():
-            pred_teacher = self.teacher_model(x_t_teacher, t_teacher)
-        
-        # Student prediction at low resolution
-        x_t_student = self._add_noise(x_0, t_student, noise)
-        pred_student = self.student_model(x_t_student, t_student)
-        
-        # Consistency loss: make student predictions close to teacher
-        loss = F.mse_loss(pred_student, pred_teacher.detach())
-        
-        return loss
-    
-    def _add_noise(self, x_0: torch.Tensor, t: torch.Tensor, noise: torch.Tensor) -> torch.Tensor:
-        """Add noise according to diffusion schedule"""
-        alpha_cumprod = torch.ones_like(t, dtype=x_0.dtype)  # Use same dtype as input
-        for i in range(len(t)):
-            alpha_cumprod[i] = 0.5 ** (t[i].to(x_0.dtype) / self.teacher_steps)  # Convert to same dtype
-
-        alpha_cumprod = alpha_cumprod.view(-1, 1, 1, 1, 1)
-        sqrt_alpha = torch.sqrt(alpha_cumprod)
-        sqrt_one_minus_alpha = torch.sqrt(1.0 - alpha_cumprod)
-
-        return sqrt_alpha * x_0 + sqrt_one_minus_alpha * noise
-    
-    def progressive_distillation(self, dataloader: DataLoader, num_distillation_steps: int = 10) -> Dict[str, float]:
-        """Compute progressive distillation losses (no optimization - caller handles training)"""
-        step_counts = self.diffusion_config.progressive_distillation
-        device = next(self.student_model.parameters()).device
-
-        distillation_results = {}
-
-        for target_steps in step_counts:
-            print(f"Computing loss for {target_steps} steps...")
-            self.student_steps = target_steps
-
-            # Loss tracking
-            total_loss = 0.0
-            num_batches = 0
-
-            for batch in tqdm(dataloader, desc=f"Computing loss {target_steps} steps"):
-                x_0 = batch['latent'].to(device)
-
-                # Sample random timesteps
-                t_student = torch.randint(0, target_steps, (x_0.shape[0],), device=device)
-                t_teacher = torch.randint(0, self.teacher_steps, (x_0.shape[0],), device=device)
-
-                # Compute consistency loss
-                loss = self.consistency_loss(x_0, t_student, t_teacher)
-
-                total_loss += loss.item()
-                num_batches += 1
-
-                if num_batches >= num_distillation_steps:
-                    break
-
-            avg_loss = total_loss / max(1, num_batches)
-            distillation_results[f'steps_{target_steps}'] = avg_loss
-            print(f"Loss for {target_steps} steps: {avg_loss:.6f}")
-
-        return distillation_results
-    
-    def fast_inference(self, shape: Tuple[int, ...], num_steps: int = 4) -> torch.Tensor:
-        """Fast 4-step inference using student model"""
-        # Get device and dtype from model parameters
-        device = next(self.student_model.parameters()).device
-        dtype = next(self.student_model.parameters()).dtype
-
-        # Initialize with random noise
-        x_t = torch.randn(shape, device=device, dtype=dtype)
-
-        # Progressive denoising in 4 steps
-        step_size = self.diffusion_config.timesteps // num_steps
-
-        for i in range(num_steps):
-            # Create timestep tensor
-            current_step = self.diffusion_config.timesteps - i * step_size - 1
-            t = torch.full((shape[0],), current_step, device=device, dtype=dtype)
-
-            # Predict noise using student model
-            pred_noise = self.student_model(x_t, t)
-
-            # Remove noise using simplified DDIM step
-            # Calculate alpha_t = alpha_t^2 (since we're denoising from noise to signal)
-            alpha_t = torch.pow(torch.tensor(0.5, device=device, dtype=torch.float32), (current_step / self.diffusion_config.timesteps))
-            alpha_t = alpha_t.to(dtype)
-
-            # DDIM update: x_{t-1} = sqrt(alpha_{t-1}) * (x_t - sqrt(1-alpha_t) * pred_noise) / sqrt(alpha_t)
-            sqrt_alpha_t = torch.sqrt(alpha_t + 1e-8)
-            sqrt_one_minus_alpha_t = torch.sqrt(1 - alpha_t + 1e-8)
-
-            # Simplified update: x_{t-1} = (x_t - (1 - alpha_t) * pred_noise) / sqrt(alpha_t)
-            coeff = 1 - alpha_t
-            x_t = (x_t - coeff * pred_noise) / sqrt_alpha_t
-
-        return x_t
-
-# ============================================================================
-# NOISE SCHEDULING & DIFFUSION UTILITIES
-# ============================================================================
-
-class NoiseSchedule:
-    """Linear noise schedule for diffusion with consistency support"""
-    
-    def __init__(self, config: DiffusionConfig):
-        self.timesteps = config.timesteps
-        self.betas = torch.linspace(config.beta_start, config.beta_end, self.timesteps)
-        self.alphas = 1.0 - self.betas
-        self.alphas_cumprod = torch.cumprod(self.alphas, dim=0)
-        self.alphas_cumprod_prev = F.pad(self.alphas_cumprod[:-1], (1, 0), value=1.0)
-        self.sqrt_alphas_cumprod = torch.sqrt(self.alphas_cumprod)
-        self.sqrt_one_minus_alphas_cumprod = torch.sqrt(1.0 - self.alphas_cumprod)
-        self.sqrt_recip_alphas_cumprod = torch.sqrt(1.0 / self.alphas_cumprod)
-        self.sqrt_recipm1_alphas_cumprod = torch.sqrt(1.0 / self.alphas_cumprod - 1.0)
-    
-    def q_sample(self, x_0: torch.Tensor, t: torch.Tensor, noise: torch.Tensor) -> torch.Tensor:
-        """Forward diffusion process: x_t = sqrt(alpha_cumprod_t) * x_0 + sqrt(1 - alpha_cumprod_t) * noise"""
-        sqrt_alpha = self.sqrt_alphas_cumprod[t].view(-1, 1, 1, 1, 1)
-        sqrt_one_minus_alpha = self.sqrt_one_minus_alphas_cumprod[t].view(-1, 1, 1, 1, 1)
-        return sqrt_alpha * x_0 + sqrt_one_minus_alpha * noise
-    
-    def to(self, device, dtype=None):
-        self.betas = self.betas.to(device, dtype=dtype if dtype is not None else self.betas.dtype)
-        self.alphas = self.alphas.to(device, dtype=dtype if dtype is not None else self.alphas.dtype)
-        self.alphas_cumprod = self.alphas_cumprod.to(device, dtype=dtype if dtype is not None else self.alphas_cumprod.dtype)
-        self.alphas_cumprod_prev = self.alphas_cumprod_prev.to(device, dtype=dtype if dtype is not None else self.alphas_cumprod_prev.dtype)
-        self.sqrt_alphas_cumprod = self.sqrt_alphas_cumprod.to(device, dtype=dtype if dtype is not None else self.sqrt_alphas_cumprod.dtype)
-        self.sqrt_one_minus_alphas_cumprod = self.sqrt_one_minus_alphas_cumprod.to(device, dtype=dtype if dtype is not None else self.sqrt_one_minus_alphas_cumprod.dtype)
-        self.sqrt_recip_alphas_cumprod = self.sqrt_recip_alphas_cumprod.to(device, dtype=dtype if dtype is not None else self.sqrt_recip_alphas_cumprod.dtype)
-        self.sqrt_recipm1_alphas_cumprod = self.sqrt_recipm1_alphas_cumprod.to(device, dtype=dtype if dtype is not None else self.sqrt_recipm1_alphas_cumprod.dtype)
-        return self
-
-# ============================================================================
-# ARCHITECTURE: LATENT DIFFUSION + 3D CONVERTER WITH MEMORY OPTIMIZATIONS
-# ============================================================================
-
-class SpatialAttention(nn.Module):
-    """Self-attention for spatial feature maps with grouped-query attention"""
-    
-    def __init__(self, channels: int, num_heads: int = 8, num_groups: int = 4):
-        super().__init__()
-        self.num_heads = num_heads
-        self.num_groups = num_groups
-        self.channels = channels
-        self.scale = (channels // num_heads) ** -0.5
-        
-        # Use grouped-query attention instead of multi-head
-        self.grouped_attention = GroupedQueryAttention(channels, num_groups, num_groups)
-    
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return self.grouped_attention(x)
-
-class ResidualBlock3D(nn.Module):
-    """3D residual block with optional attention and gradient checkpointing"""
-    
-    def __init__(self, in_channels: int, out_channels: int, time_emb_dim: int, 
-                 use_attention: bool = False, enable_checkpointing: bool = True):
-        super().__init__()
-        
-        self.time_mlp = nn.Sequential(
-            nn.Linear(time_emb_dim, out_channels),
-            nn.SiLU(),
-            nn.Linear(out_channels, out_channels)
-        )
-        
-        self.block1 = nn.Sequential(
-            nn.InstanceNorm3d(in_channels),
-            nn.SiLU(),
-            nn.Conv3d(in_channels, out_channels, 3, padding=1)
-        )
-
-        self.block2 = nn.Sequential(
-            nn.InstanceNorm3d(out_channels),
-            nn.SiLU(),
-            nn.Conv3d(out_channels, out_channels, 3, padding=1)
-        )
-
-        self.out_channels = out_channels
-        
-        self.res_conv = nn.Conv3d(in_channels, out_channels, 1) if in_channels != out_channels else nn.Identity()
-        
-        # Use grouped-query attention with memory optimization
-        if use_attention:
-            self.attention = SpatialAttention(out_channels, num_groups=4)
-        else:
-            self.attention = nn.Identity()
-        
-        # Apply gradient checkpointing wrapper
-        if enable_checkpointing:
-            self.block1 = GradientCheckpointingWrapper(self.block1)
-            self.block2 = GradientCheckpointingWrapper(self.block2)
-    
-    def forward(self, x: torch.Tensor, time_emb: torch.Tensor) -> torch.Tensor:
-        h = self.block1(x)
-        h = h + self.time_mlp(time_emb).view(-1, self.out_channels, 1, 1, 1)
-        h = self.block2(h)
-        h = h + self.res_conv(x)
-        h = self.attention(h)
-        return h
-
-class LatentDiffusionUNet(nn.Module):
-    """UNet for diffusion on latent codes with memory optimizations"""
-
-    def __init__(self, config: ModelConfig, diffusion_config: DiffusionConfig):
-        super().__init__()
-        self.latent_dim = config.latent_dim
-        self.diffusion_config = diffusion_config
-        self.encoder_out_dim = config.encoder_channels[0] * 2 * 2 * 2  # Reduced from 4x4x4 to 2x2x2 to avoid overflow
-        self.config = config
-
-        time_emb_dim = config.latent_dim
-        self.time_embedding = nn.Sequential(
-            nn.Linear(1, time_emb_dim),
-            nn.SiLU(),
-            nn.Linear(time_emb_dim, time_emb_dim)
-        )
-
-        # Encoder: project latent to spatial
-        self.encoder = nn.Sequential(
-            nn.Linear(config.latent_dim, self.encoder_out_dim),
-            nn.SiLU(),
-            nn.Linear(self.encoder_out_dim, self.encoder_out_dim),
-        )
-
-        channels = config.encoder_channels + [config.decoder_channels[-1]]
-        self.down_blocks = nn.ModuleList()
-        self.down_convs = nn.ModuleList()
-
-        for i in range(len(channels) - 1):
-            self.down_blocks.append(ResidualBlock3D(
-                channels[i], channels[i+1], time_emb_dim, 
-                use_attention=False, 
-                enable_checkpointing=config.enable_gradient_checkpointing
-            ))
-            self.down_convs.append(nn.Conv3d(channels[i+1], channels[i+1], 3, stride=1, padding=1))
-
-        self.mid_block = ResidualBlock3D(
-            channels[-1], channels[-1], time_emb_dim, 
-            use_attention=False,
-            enable_checkpointing=config.enable_gradient_checkpointing
-        )
-
-        self.up_convs = nn.ModuleList()
-        self.up_blocks = nn.ModuleList()
-        for i in range(len(channels) - 1, 0, -1):
-            self.up_convs.append(nn.Conv3d(channels[i], channels[i-1], 3, stride=1, padding=1))
-            self.up_blocks.append(ResidualBlock3D(
-                channels[i-1], channels[i-1], time_emb_dim, 
-                use_attention=False,
-                enable_checkpointing=config.enable_gradient_checkpointing
-            ))
-
-        self.out_conv = nn.Conv3d(channels[0], channels[0], 1)
-        self.out = nn.Linear(self.encoder_out_dim, self.latent_dim)
-        
-        # Apply torch.compile for kernel fusion
-        if config.use_torch_compile:
-            self._apply_torch_compile()
-    
-    def _apply_torch_compile(self):
-        """Apply torch.compile() with reduce-overhead mode for kernel fusion"""
-        # Check if torch.compile is enabled in config
-
-        # Try different backends in order of preference to handle Triton issues
-        backends_to_try = [
-            ("inductor", "reduce-overhead"),
-            ("inductor", "default"),
-            ("eager", "reduce-overhead"),
-            ("eager", "default")
-        ]
-        import traceback
-        for backend, mode in backends_to_try:
-            try:
-                print(f"Trying torch.compile with backend='{backend}', mode='{mode}'...")
-
-                if backend == "inductor":
-                    # Try to configure inductor to avoid Triton issues
-                    import torch._inductor.config
-                    if hasattr(torch._inductor.config, 'triton'):
-                        triton_config = torch._inductor.config.triton
-                        if hasattr(triton_config, 'cudagraphs'):
-                            triton_config.cudagraphs = False
-                        # autotune doesn't exist in this PyTorch version, skip it
-                    else:
-                        print("⚠️ Triton config not available, using default inductor settings")
-
-                # Try to compile
-                self.forward = torch.compile(self.forward, backend=backend, mode=mode)
-                print(f"✅ Successfully applied torch.compile() with backend='{backend}', mode='{mode}'")
-                return
-
-            except Exception as e:
-                print(f"❌ torch.compile() failed with backend='{backend}': {str(e)}")
-                traceback.print_exc()
-                continue
-
-        print("⚠️  All torch.compile() backends failed, using original forward function")
-        # Keep original forward function - no functionality lost
-        pass
-    
-    def forward(self, x: torch.Tensor, timestep: torch.Tensor, condition: torch.Tensor = None) -> torch.Tensor:
-        """
-        Forward pass with memory optimizations.
-        x: [B, latent_dim] - noisy latent codes
-        timestep: [B] - diffusion timesteps
-        condition: [B, C, D, H, W] - optional spatial conditioning
-        """
-        b = x.shape[0]
-        
-        t_emb = self.time_embedding(timestep.to(self.time_embedding[0].weight.dtype).unsqueeze(1) / self.diffusion_config.timesteps)
-        
-        # Expand latent to 3D spatial (2x2x2)
-        h = self.encoder(x)
-        h = h.view(b, -1)
-        target_size = self.encoder_out_dim
-        if h.size(1) > target_size:
-            h = h[:, :target_size]
-        elif h.size(1) < target_size:
-            h = torch.cat([h, h.new_zeros(b, target_size - h.size(1))], dim=1)
-        h = h.view(b, self.config.encoder_channels[0], 2, 2, 2)
-        
-        if condition is not None and condition.shape == h.shape:
-            h = h + condition
-        
-        # U-Net forward pass
-        skip_connections = []
-        for i in range(len(self.down_blocks)):
-            h = self.down_blocks[i](h, t_emb)
-            h = self.down_convs[i](h)
-            skip_connections.append(h)
-
-        h = self.mid_block(h, t_emb)
-
-        for i in range(len(self.up_blocks)):
-            skip = skip_connections.pop()
-            h = h + skip
-            h = self.up_convs[i](h)
-            h = self.up_blocks[i](h, t_emb)
-        
-        out = self.out_conv(h).view(b, -1)
-        out = self.out(out)
-        return out
-
-class LatentTo3DConverter(nn.Module):
-    """Convert n-dimensional latent codes to 3D spatial representation with adaptive output"""
-
-    def __init__(self, latent_dim: int, grid_resolution: int = 32):
-        super().__init__()
-        self.latent_dim = latent_dim
-        self.grid_resolution = grid_resolution
-        self.output_shape = (grid_resolution, grid_resolution, grid_resolution)
-        total_voxels = grid_resolution ** 3
-
-        self.decoder = nn.Sequential(
-            nn.Linear(latent_dim, 1024),
-            nn.ReLU(),
-            nn.Linear(1024, 2048),
-            nn.ReLU(),
-            nn.Linear(2048, total_voxels)
-        )
-    
-    def set_resolution(self, new_resolution: int):
-        """Update working output resolution for interpolation"""
-        if new_resolution == self.grid_resolution:
-            return
-        self.grid_resolution = new_resolution
-        print(f"LatentTo3DConverter: Target resolution set to {new_resolution}^3")
-
-    def forward(self, latent: torch.Tensor, target_res: int = None) -> torch.Tensor:
-        """Convert latent code to voxel grid, optionally interpolating to target_res"""
-        batch_size = latent.shape[0]
-        voxels = self.decoder(latent)
-
-        # Original decoder output size was fixed at init
-        base_res = int(round(voxels.shape[1]**(1/3)))
-        voxels = voxels.view(batch_size, base_res, base_res, base_res)
-
-        target = target_res if target_res is not None else self.grid_resolution
-
-        if target != base_res:
-            voxels = F.interpolate(
-                voxels.unsqueeze(1),
-                size=(target, target, target),
-                mode='trilinear',
-                align_corners=False
-            ).squeeze(1)
-
-        return voxels
-
-# ============================================================================
-# PIPELINE PARALLELISM: CFD + DIFFUSION OVERLAP
-# ============================================================================
-
-class PipelineParallelism:
-    """Pipeline parallelism to overlap CFD computation with diffusion sampling"""
-    
-    def __init__(self, config: TrainingConfig):
-        self.config = config
-        self.num_stages = config.num_pipeline_stages
-        self.enable_overlap = config.enable_pipeline_parallelism
-    
-    async def pipeline_process(self, diffusion_model, cfd_solver, batch_data: Dict[str, torch.Tensor]):
-        """
-        Overlap diffusion and CFD computations in pipeline.
-        
-        Stage 1: Diffusion sampling
-        Stage 2: CFD simulation
-        """
-        device = next(diffusion_model.parameters()).device
-        
-        if not self.enable_overlap:
-            # Sequential processing
-            with torch.no_grad():
-                latent_sample = diffusion_model.sample(batch_data['latent'].to(device))
-                voxel_grid = self._convert_to_voxel_grid(latent_sample)
-                cfd_results = await self._run_cfd_async(cfd_solver, voxel_grid)
-            return cfd_results
-        
-        # Pipeline parallelism
-        results = []
-        
-        # Create pipeline tasks
-        tasks = []
-        for i in range(batch_data['latent'].shape[0]):
-            task = self._pipeline_stage(diffusion_model, cfd_solver, batch_data['latent'][i:i+1])
-            tasks.append(task)
-        
-        # Execute pipeline
-        results = await asyncio.gather(*tasks)
-        return results
-    
-    async def _pipeline_stage(self, diffusion_model, cfd_solver, sample: torch.Tensor):
-        """Single pipeline stage combining diffusion and CFD"""
-        device = next(diffusion_model.parameters()).device
-        sample = sample.to(device)
-        
-        # Stage 1: Fast diffusion sampling (4 steps)
-        with torch.no_grad():
-            latent_sample = self._fast_diffusion_sampling(diffusion_model, sample)
-            voxel_grid = self._convert_to_voxel_grid(latent_sample)
-        
-        # Stage 2: Parallel CFD simulation
-        cfd_results = await self._run_cfd_async(cfd_solver, voxel_grid)
-        
-        return {
-            'latent': latent_sample.cpu(),
-            'voxel_grid': voxel_grid.cpu(),
-            'cfd_results': cfd_results
-        }
-    
-    def _fast_diffusion_sampling(self, diffusion_model: ConsistencyModel, sample: torch.Tensor) -> torch.Tensor:
-        """Fast 4-step diffusion sampling using student model"""
-        return diffusion_model.student_model.fast_inference(sample.shape, num_steps=4)
-    
-    def _convert_to_voxel_grid(self, latent: torch.Tensor) -> torch.Tensor:
-        """Convert latent sample to voxel grid"""
-        # Simple conversion for pipeline testing
-        return torch.sigmoid(latent).view(1, 32, 32, 32)
-    
-    async def _run_cfd_async(self, cfd_solver: D3Q27CascadedSolver, voxel_grid: torch.Tensor) -> Dict[str, float]:
-        """Run CFD simulation asynchronously"""
-        # Convert voxel grid to geometry mask
-        geometry_mask = (voxel_grid > 0.5).float()
-        
-        # Run LBM solver
-        cfd_solver.collide_stream(geometry_mask, steps=100)
-        
-        # Compute aerodynamic coefficients
-        results = cfd_solver.compute_aerodynamic_coefficients(geometry_mask)
-        
-        return results
-
-# ============================================================================
-# ENHANCED CFD SIMULATION WITH FLUIDX3D INTEGRATION
-# ============================================================================
-
-class AdvancedCFDSimulator:
-    """Advanced CFD simulator with FluidX3D integration and adaptive mesh refinement"""
-    
-    def __init__(self, config: CFDConfig, device: torch.device):
-        self.config = config
-        self.device = device
-        self.resolution = config.base_grid_resolution
-
-        if config.solver_type != "D3Q27":
-            raise ValueError("Only the D3Q27 LBM solver is supported")
-        self.lbm_solver = D3Q27CascadedSolver(self.config, device, LBMPhysicsConfig)
-
-        # Initialize a higher-resolution solver for AMR if enabled
-        if self.config.use_amr:
-            import copy
-            amr_config = copy.deepcopy(self.config)
-            amr_config.resolution = self.config.base_grid_resolution * 2  # Double the resolution
-            self.amr_solver = D3Q27CascadedSolver(amr_config, device, LBMPhysicsConfig)
-        else:
-            self.amr_solver = None
-
-        # Initialize flow field
-        self.init_flow_field()
-    
-    def set_resolution(self, resolution: int):
-        """Update simulator resolution and re-initialize solver"""
-        if resolution == self.resolution:
-            return
-
-        self.resolution = resolution
-        self.config.base_grid_resolution = resolution
-
-        # Re-initialize LBM solver with new resolution
-        self.lbm_solver = D3Q27CascadedSolver(self.config, self.device, LBMPhysicsConfig)
-
-        if self.config.use_amr:
-            import copy
-            amr_config = copy.deepcopy(self.config)
-            amr_config.resolution = resolution * 2
-            self.amr_solver = D3Q27CascadedSolver(amr_config, self.device, LBMPhysicsConfig)
-        else:
-            self.amr_solver = None
-
-        self.init_flow_field()
-        print(f"AdvancedCFDSimulator: Solver resolution updated to {resolution}^3")
-
-    def init_flow_field(self):
-        """Initialize flow field for incompressible flow"""
-        # Initialize LBM solver
-        self.lbm_solver._initialize_equilibrium()
-        if self.amr_solver:
-            self.amr_solver._initialize_equilibrium()
-    
-    def simulate_aerodynamics(self, geometry: torch.Tensor, steps: int = 100) -> Dict[str, float]:
-        """
-        Simulate flow around geometry with adaptive mesh refinement.
-        geometry: [D, H, W] binary voxel grid (1 = solid, 0 = fluid)
-        """
-        device = geometry.device
-
-        # Step 1: Run the base solver
-        geometry_mask = (geometry > 0.5).float()
-        print(f"Running {self.config.solver_type} GPU LBM solver at base resolution...")
-        self.lbm_solver.collide_stream(geometry_mask, steps=steps)
-        results = self.lbm_solver.compute_aerodynamic_coefficients(geometry_mask)
-
-        # Step 2: If AMR is enabled, run the high-resolution solver
-        if self.amr_solver:
-            print("Applying adaptive mesh refinement by running a higher-resolution simulation...")
-
-            # Upsample the geometry for the AMR solver
-            amr_geometry = F.interpolate(
-                geometry.unsqueeze(0).unsqueeze(0),
-                size=(self.amr_solver.resolution, self.amr_solver.resolution, self.amr_solver.resolution),
-                mode='nearest'
-            ).squeeze(0).squeeze(0)
-            amr_geometry_mask = (amr_geometry > 0.5).float()
-
-            self.amr_solver.collide_stream(amr_geometry_mask, steps=steps)
-            amr_results = self.amr_solver.compute_aerodynamic_coefficients(amr_geometry_mask)
-
-            # Blend the results for a more accurate final value
-            results['drag_coefficient'] = (results['drag_coefficient'] + amr_results['drag_coefficient']) / 2
-            results['lift_coefficient'] = (results['lift_coefficient'] + amr_results['lift_coefficient']) / 2
-
-        # Step 3: Run FluidX3D for validation (if available)
-        fluidx3d_results = self._run_fluidx3d_validation(geometry)
-        if fluidx3d_results:
-            # Blend results for accuracy
-            results['drag_coefficient'] = 0.7 * results['drag_coefficient'] + 0.3 * fluidx3d_results['drag_coefficient']
-            results['lift_coefficient'] = 0.7 * results['lift_coefficient'] + 0.3 * fluidx3d_results['lift_coefficient']
-
-        return results
-    
-    def _run_fluidx3d_validation(self, voxel_grid: torch.Tensor) -> Optional[Dict[str, float]]:
-        """Run FluidX3D for validation (simplified integration)"""
-        try:
-            # Convert to STL and run FluidX3D
-            stl_path = self._voxel_to_stl_path(voxel_grid)
-            if stl_path and os.path.exists(stl_path):
-                # Run simplified FluidX3D simulation
-                return self._run_fluidx3d_fast(stl_path)
-        except Exception as e:
-            print(f"FluidX3D validation failed: {e}")
-        
-        return None
-    
-    def _voxel_to_stl_path(self, voxel_grid: torch.Tensor) -> Optional[str]:
-        """Convert voxel grid to STL file path"""
-        try:
-            voxel_np = voxel_grid.cpu().numpy()
-            binary_grid = (voxel_np > 0.5).astype(np.float32)
-            vertices, faces, _, _ = measure.marching_cubes(
-                binary_grid,
-                level=0.5,
-                spacing=(1.0, 1.0, 1.0)
-            )
-            # Match the same centered physical frame used by the internal solver
-            # and the OpenFOAM validation case: unit cube centered at the origin.
-            scale = float(self.config.lbm_config.physical_length_scale)
-            h = scale / float(self.config.base_grid_resolution)
-            vertices = vertices * h - (scale * 0.5) + (0.5 * h)
-
-            # Create mesh
-            mesh = trimesh.Trimesh(vertices=vertices, faces=faces)
-            with tempfile.NamedTemporaryFile(suffix='.stl', delete=False) as tmp:
-                mesh.export(tmp.name)
-                return tmp.name
-        except Exception as e:
-            print(f"STL conversion failed: {e}")
-            return None
-    
-    def _run_fluidx3d_fast(self, stl_path: str) -> Dict[str, float]:
-        """Run FluidX3D with fast settings"""
-        # Simplified FluidX3D integration
-        # This would use the actual FluidX3D executable in a real implementation
-        
-        # For now, return physics-based approximation
-        volume = 0.1  # Approximate volume fraction
-        return {
-            'drag_coefficient': 0.02 + volume * 0.1,
-            'lift_coefficient': volume * 0.4
-        }
-
-# ============================================================================
-# DATASET & DATA LOADING
-# ============================================================================
-import random
-class AircraftDesignDataset(Dataset):
-    """Synthetic dataset for aircraft structure training with adaptive resolution"""
-
-    def __init__(self, num_samples: int = 10000, grid_size: int = 32, seed: int = None, latent_dim: int = 128, target_grid_size: int = 1024):
-        self.num_samples = num_samples
-        self.grid_size = grid_size
-        self.target_grid_size = target_grid_size
-
-        if seed is None:
-            seed = random.randint(0, 1000000)
-        np.random.seed(seed)
-        torch.manual_seed(seed)
-
-        self.latent_codes = torch.randn(num_samples, latent_dim)
-        self.geometries = self._generate_geometries()
-    
-    def set_resolution(self, new_grid_size: int):
-        """Update dataset resolution for progressive training via interpolation"""
-        if new_grid_size == self.grid_size:
-            return
-
-        print(f"Dataset: Interpolating geometries from {self.grid_size}^3 to {new_grid_size}^3...")
-        new_geometries = []
-        for geom in self.geometries:
-            geom_reshaped = geom.unsqueeze(0).unsqueeze(0)
-            interpolated = F.interpolate(
-                geom_reshaped,
-                size=(new_grid_size, new_grid_size, new_grid_size),
-                mode='trilinear',
-                align_corners=False
-            )
-            new_geometries.append((interpolated.squeeze() > 0.5).float())
-
-        self.geometries = new_geometries
-        self.grid_size = new_grid_size
-
-    def _generate_geometries(self) -> List[torch.Tensor]:
-        """Generate synthetic aircraft geometries using vectorized ops"""
-        geometries = []
-        for i in range(self.num_samples):
-            geom = torch.zeros(self.grid_size, self.grid_size, self.grid_size)
-            scale = self.grid_size / 32.0
-            cx, cy, cz = self.grid_size // 2, self.grid_size // 2, self.grid_size // 2
-            
-            z_indices, y_indices, x_indices = torch.meshgrid(
-                torch.arange(self.grid_size),
-                torch.arange(self.grid_size),
-                torch.arange(self.grid_size),
-                indexing='ij'
-            )
-
-            dist_center = torch.sqrt((x_indices - cx)**2 + (z_indices - cz)**2)
-            fuselage_mask = (dist_center < 6 * scale) & (y_indices > 10 * scale) & (y_indices < 22 * scale)
-            geom[fuselage_mask] = 1.0
-
-            wing_mask = (y_indices > 8 * scale) & (y_indices < 24 * scale) & \
-                        ((z_indices < 4 * scale) | (z_indices > self.grid_size - 4 * scale))
-            geom[wing_mask] = 1.0
-            
-            noise = torch.rand_like(geom)
-            geom = (geom + 0.1 * noise > 0.5).float()
-            geometries.append(geom)
-        
-        return geometries
-    
-    def __len__(self) -> int:
-        return self.num_samples
-    
-    def __getitem__(self, idx: int) -> Dict[str, torch.Tensor]:
-        return {
-            'latent': self.latent_codes[idx],
-            'geometry': self.geometries[idx],
-            'target_speed': torch.tensor(self.grid_size / 32 * 50.0)
-        }
-
-# ============================================================================
-# LOSS FUNCTIONS
-# ============================================================================
-
-class ConnectivityLoss(nn.Module):
-    """Penalize disconnected voxel groups"""
-    
-    def __init__(self, penalty: float = 10.0):
-        super().__init__()
-        self.penalty = penalty
-    
-    def forward(self, voxel_grid: torch.Tensor) -> torch.Tensor:
-        """Compute connectivity penalty for batch of voxel grids"""
-        batch_size = voxel_grid.shape[0]
-        total_penalty = 0.0
-        
-        for b in range(batch_size):
-            geom = (voxel_grid[b] > 0.5).int().cpu().numpy()
-            
-            # Label connected components
-            labeled, num_components = label(geom)
-            
-            if num_components > 1:
-                # Count voxels in each component
-                component_sizes = np.bincount(labeled.flatten())
-                
-                # Largest component should dominate
-                largest_size = component_sizes[1:].max() if num_components > 1 else 0
-                total_size = geom.sum()
-                
-                if largest_size > 0:
-                    disconnected_fraction = (total_size - largest_size) / (total_size + 1e-6)
-                    total_penalty += disconnected_fraction
-        
-        result = self.penalty * total_penalty / batch_size if batch_size > 0 else 0.0
-        return torch.tensor(result, device=voxel_grid.device, dtype=torch.float32)
-
-class AerodynamicLoss(nn.Module):
-    """Loss based on aerodynamic properties using advanced CFD"""
-
-    def __init__(self):
-        super().__init__()
-
-    def forward(self, voxel_grid: torch.Tensor, design_spec: DesignSpec, cfd_simulator: "AdvancedCFDSimulator") -> torch.Tensor:
-        """
-        Compute aerodynamic loss balancing drag, lift, and volume using advanced CFD.
-        """
-        batch_size = voxel_grid.shape[0]
-        loss = torch.tensor(0.0, device=voxel_grid.device)
-
-        for b in range(batch_size):
-            # Get single voxel grid for CFD
-            single_voxel_grid = voxel_grid[b]
-            geometry = (single_voxel_grid > 0.5).float()
-
-            # Run advanced CFD analysis with the provided simulator
-            cfd_results = cfd_simulator.simulate_aerodynamics(geometry, steps=100)
-
-            # Volume penalty (space weight)
-            volume = geometry.sum() / np.prod(geometry.shape)
-            volume_loss = design_spec.space_weight * volume
-
-            # Drag coefficient penalty (drag weight)
-            cd = cfd_results.get('drag_coefficient', 0.1)
-            drag_loss = design_spec.drag_weight * cd
-
-            # Lift coefficient encouragement (lift weight)
-            cl = abs(cfd_results.get('lift_coefficient', 0.0))
-            lift_loss = design_spec.lift_weight * (1.0 - torch.clamp(torch.tensor(cl, device=voxel_grid.device), 0, 1))
-
-            loss += volume_loss + drag_loss + lift_loss
-
-        return loss / batch_size
-
-# ============================================================================
-# TRAINING PIPELINE WITH ALL OPTIMIZATIONS
-# ============================================================================
-
-class OptimizedDiffusionTrainer:
-    """Main training orchestrator with all TRM/HRM optimizations"""
-    
-    def __init__(
-        self,
-        model_config: ModelConfig,
-        diffusion_config: DiffusionConfig,
-        training_config: TrainingConfig,
-        cfd_config: CFDConfig,
-        device: torch.device = None
-    ):
-        self.device = device or torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-        
-        self.model_config = model_config
-        self.diffusion_config = diffusion_config
-        self.training_config = training_config
-        self.cfd_config = cfd_config
-        
-        # Precision handling for mixed precision training
-        self.precision_dtypes = {
-            'float64': torch.float64,
-            'double': torch.float64,
-            'float32': torch.float32,
-            'float': torch.float32,
-            'float16': torch.float16,
-            'half': torch.float16,
-            'bfloat16': torch.bfloat16,
-            'float8': torch.float8 if hasattr(torch, 'float8') else torch.float16
-        }
-        self.dtype = self.precision_dtypes.get(training_config.precision, torch.float32)
-        print(f"Using precision: {training_config.precision} ({self.dtype})")
-
-        self.noise_schedule = NoiseSchedule(diffusion_config).to(self.device, self.dtype)
-
-        # Models with optimizations
-        self.diffusion_model = LatentDiffusionUNet(model_config, diffusion_config).to(self.device).to(self.dtype)
-        self.converter = LatentTo3DConverter(model_config.latent_dim, model_config.grid_resolution).to(self.device).to(self.dtype)
-        
-        # Progressive growth state
-        self.current_grid_size = model_config.grid_resolution
-
-        # 4-step consistency model
-        self.consistency_model = ConsistencyModel(model_config, diffusion_config, self.dtype).to(self.device)
-        
-        # Initialize EMA model
-        self.ema_model = self._copy_model(self.diffusion_model)
-        
-        # Optimizer
-        params = (list(self.diffusion_model.parameters()) + 
-                 list(self.converter.parameters()) +
-                 list(self.consistency_model.student_model.parameters()))
-        self.optimizer = AdamW(params, lr=training_config.learning_rate, weight_decay=training_config.weight_decay)
-        self.scheduler = CosineAnnealingLR(self.optimizer, T_max=training_config.num_epochs)
-        
-        # Gradient scaler for mixed precision
-        self.scaler = GradScaler()
-        
-        # Losses
-        self.mse_loss = nn.MSELoss()
-        self.connectivity_loss = ConnectivityLoss(penalty=training_config.disconnection_penalty)
-        self.aero_loss = AerodynamicLoss()
-        
-        # Advanced CFD simulator for training (fast, coarse)
-        self.cfd_simulator = AdvancedCFDSimulator(cfd_config, self.device)
-        
-        # High-fidelity CFD simulator for validation (accurate, refined)
-        import copy
-        val_cfd_config = copy.deepcopy(cfd_config)
-        val_cfd_config.solver_type = "D3Q27"
-        val_cfd_config.use_amr = True  # Enable AMR for validation
-        self.val_cfd_simulator = AdvancedCFDSimulator(val_cfd_config, self.device)
-
-        # Pipeline parallelism
-        self.pipeline = PipelineParallelism(training_config)
-        
-        # Logging
-        self.writer = SummaryWriter(log_dir='./runs')
-        self.global_step = 0
-    
-    def _copy_model(self, model: nn.Module) -> nn.Module:
-        """Create an independent copy of the model"""
-        import copy
-        return copy.deepcopy(model)
-    
-    def _update_ema(self):
-        """Update exponential moving average model"""
-        decay = self.training_config.ema_decay
-        for ema_param, param in zip(self.ema_model.parameters(), self.diffusion_model.parameters()):
-            ema_param.data.mul_(decay).add_(param.data, alpha=1 - decay)
-
-    def validate_epoch(self, val_loader: DataLoader, grid_size: int = 32) -> Dict[str, float]:
-        """Validate for one epoch with the high-fidelity D3Q27 solver"""
-        # Synchronize resolutions
-        if self.current_grid_size != grid_size:
-            self.converter.set_resolution(grid_size)
-            self.cfd_simulator.set_resolution(grid_size)
-            self.current_grid_size = grid_size
-
-        self.diffusion_model.eval()
-        self.converter.eval()
-
-        total_aero_loss = 0.0
-
-        pbar = tqdm(val_loader, desc=f"Validating with D3Q27 solver (grid={grid_size}x{grid_size}x{grid_size})")
-
-        with torch.no_grad():
-            for batch_idx, batch in enumerate(pbar):
-                latent = batch['latent'].to(self.device, dtype=self.dtype)
-
-                # Generate a design using the consistency model
-                generated_latent = self.consistency_model.fast_inference(latent.shape, num_steps=4)
-                voxel_grid = self.converter(generated_latent).nan_to_num(0.0)
-                voxel_grid = torch.sigmoid(voxel_grid).nan_to_num(0.0)
-
-                # CFD-based aerodynamic loss with the D3Q27 solver
-                design_spec = DesignSpec(target_speed=50.0)
-                aero_loss_val = self.aero_loss(voxel_grid, design_spec, self.val_cfd_simulator).nan_to_num(0.0)
-
-                total_aero_loss += aero_loss_val.item()
-
-        avg_aero_loss = total_aero_loss / len(val_loader)
-
-        self.writer.add_scalar('Loss/val_aerodynamic', avg_aero_loss, self.global_step)
-
-        print(f"Validation Aerodynamic Loss (D3Q27): {avg_aero_loss}")
-
-        return {'val_aerodynamic_loss': avg_aero_loss}
-
-    def train_epoch(self, train_loader: DataLoader, grid_size: int = 32) -> Dict[str, float]:
-        """Train for one epoch with all optimizations"""
-        # Synchronize resolutions
-        if self.current_grid_size != grid_size:
-            self.converter.set_resolution(grid_size)
-            self.cfd_simulator.set_resolution(grid_size)
-            self.current_grid_size = grid_size
-
-        self.diffusion_model.train()
-        self.converter.train()
-        self.consistency_model.student_model.train()
-        
-        total_loss = 0.0
-        total_mse = 0.0
-        total_consistency = 0.0
-        total_connectivity = 0.0
-        total_aero = 0.0
-        
-        pbar = tqdm(train_loader, desc=f"Training with optimizations (grid={grid_size}x{grid_size}x{grid_size})")
-        
-        for batch_idx, batch in enumerate(pbar):
-            latent = batch['latent'].to(self.device, dtype=self.dtype)
-            geometry_target = batch['geometry'].to(self.device, dtype=self.dtype)
-
-            # Resize geometry to current grid size
-            if grid_size != geometry_target.shape[1]:
-                geometry_target = F.interpolate(
-                    geometry_target.unsqueeze(1),
-                    size=(grid_size, grid_size, grid_size),
-                    mode='nearest'
-                ).squeeze(1)
-
-            # Convert latent to voxel grid
-            voxel_grid = self.converter(latent).nan_to_num(0.0)
-            voxel_grid = torch.sigmoid(voxel_grid).nan_to_num(0.0)
-
-            # Progressive distillation training
-            consistency_loss = torch.tensor(0.0, device=self.device)
-            if batch_idx % 20 == 0:  # Every 20 batches
-                consistency_loss = self._compute_consistency_loss(latent)
-
-            # Random timestep for diffusion training
-            t = torch.randint(0, self.diffusion_config.timesteps, (latent.shape[0],), device=self.device).nan_to_num(0.0)
-
-            # Forward diffusion
-            noise = torch.randn_like(latent)
-            noisy_latent = self.noise_schedule.q_sample(latent, t, noise).nan_to_num(0.0)
-
-            # Model prediction
-            pred_noise = self.diffusion_model(noisy_latent, t).nan_to_num(0.0)
-
-            # MSE loss
-            mse_loss_val = self.mse_loss(pred_noise, noise).nan_to_num(0.0)
-
-            # Connectivity loss
-            connectivity_loss_val = self.connectivity_loss(voxel_grid).nan_to_num(0.0)
-
-            # CFD-based aerodynamic loss (every 10 batches for speed)
-            aero_loss_val = torch.tensor(0.0, device=self.device)
-            if batch_idx % 10 == 0:
-                # Target speed scales with resolution for physical consistency
-                design_spec = DesignSpec(target_speed=grid_size / 32.0 * 50.0)
-                aero_loss_val = self.aero_loss(voxel_grid[:1], design_spec, self.cfd_simulator).nan_to_num(0.0)
-
-            # Combined loss
-            total_loss_val = mse_loss_val + consistency_loss + connectivity_loss_val + aero_loss_val
-
-            # Backward pass
-            self.optimizer.zero_grad()
-            total_loss_val.backward()
-
-            # Gradient clipping
-            torch.nn.utils.clip_grad_norm_(self.diffusion_model.parameters(), self.training_config.gradient_clip)
-            torch.nn.utils.clip_grad_norm_(self.converter.parameters(), self.training_config.gradient_clip)
-            torch.nn.utils.clip_grad_norm_(self.consistency_model.student_model.parameters(), self.training_config.gradient_clip)
-
-            # Optimizer step
-            self.optimizer.step()
-            
-            # EMA update
-            self._update_ema()
-            
-            # Logging
-            total_loss += total_loss_val.item()
-            total_mse += mse_loss_val.item()
-            total_consistency += consistency_loss.item()
-            total_connectivity += connectivity_loss_val.item()
-            total_aero += aero_loss_val.item()
-            
-            pbar.set_postfix({
-                'loss': total_loss_val.item(),
-                'mse': mse_loss_val.item(),
-                'consistency': consistency_loss.item(),
-                'conn': connectivity_loss_val.item(),
-                'aero': aero_loss_val.item()
-            })
-            
-            self.global_step += 1
-            
-            # Clear memory
-            if batch_idx % 10 == 0:
-                torch.cuda.empty_cache()
-        
-        avg_loss = total_loss / len(train_loader)
-        
-        # Log to tensorboard
-        self.writer.add_scalar('Loss/total', avg_loss, self.global_step)
-        self.writer.add_scalar('Loss/mse', total_mse / len(train_loader), self.global_step)
-        self.writer.add_scalar('Loss/consistency', total_consistency / len(train_loader), self.global_step)
-        self.writer.add_scalar('Loss/connectivity', total_connectivity / len(train_loader), self.global_step)
-        self.writer.add_scalar('Loss/aerodynamic', total_aero / len(train_loader), self.global_step)
-        
-        return {
-            'loss': avg_loss,
-            'mse': total_mse / len(train_loader),
-            'consistency': total_consistency / len(train_loader),
-            'connectivity': total_connectivity / len(train_loader),
-            'aerodynamic': total_aero / len(train_loader)
-        }
-    
-    def _compute_consistency_loss(self, latent: torch.Tensor) -> torch.Tensor:
-        """Compute consistency loss for progressive distillation"""
-        batch_size = latent.shape[0]
-        device = latent.device
-        
-        # Sample random timesteps for teacher and student
-        t_student = torch.randint(0, self.diffusion_config.student_steps, (batch_size,), device=device)
-        t_teacher = torch.randint(0, self.diffusion_config.teacher_steps, (batch_size,), device=device)
-        
-        # Compute consistency loss
-        return self.consistency_model.consistency_loss(latent, t_student, t_teacher)
-    
-    def train(self, train_loader: DataLoader, val_loader: DataLoader = None):
-        """Progressive training with adaptive resolutions based on VRAM"""
-        # Determine maximum resolution based on GPU memory
-        max_vram_res = get_vram_limit_resolution(max_usage=0.9)
-
-        # Schedule: Start at 32, double until max_vram_res or geometry limit
-        grid_sizes = [32]
-        curr = 64
-        while curr <= max_vram_res and curr <= 512: # Dense LBM cap at 512 for stability
-            grid_sizes.append(curr)
-            curr *= 2
-
-        print(f"📈 Progressive growth schedule: {grid_sizes}")
-
-        for grid_size in grid_sizes:
-            # Sync dataset resolution
-            if hasattr(train_loader.dataset, 'set_resolution'):
-                train_loader.dataset.set_resolution(grid_size)
-            if val_loader and hasattr(val_loader.dataset, 'set_resolution'):
-                val_loader.dataset.set_resolution(grid_size)
-
-            print(f"\n{'='*60}")
-            print(f"Training with grid size: {grid_size}x{grid_size}x{grid_size}")
-            print(f"Features: 4-step consistency, grouped-query attention, gradient checkpointing")
-            print(f"Memory: 60% VRAM savings, 50% KV-cache reduction")
-            print(f"CFD: Adaptive mesh (~5k cells), GPU LBM solver")
-            print(f"{'='*60}\n")
-
-            torch.cuda.empty_cache()
-
-            epochs = self.training_config.num_epochs if grid_size == 32 else max(1, self.training_config.num_epochs // 2)
-
-            for epoch in range(epochs):
-                print(f"\nGrid {grid_size} - Epoch {epoch + 1}/{epochs}")
-                
-                # Progressive distillation
-                if epoch % 10 == 0 and epoch > 0:
-                    print("Running progressive distillation...")
-                    self._run_progressive_distillation(train_loader)
-                
-                metrics = self.train_epoch(train_loader, grid_size=grid_size)
-
-                print(f"Epoch {epoch + 1} Metrics: {metrics}")
-
-                if val_loader and (epoch + 1) % self.training_config.val_interval == 0:
-                    self.validate_epoch(val_loader, grid_size=grid_size)
-
-                if (epoch + 1) % self.training_config.save_interval == 0:
-                    self.save_checkpoint(f'checkpoint_optimized_grid{grid_size}_ep{epoch+1}.pt')
-
-            self.scheduler.step()
-    
-    def _run_progressive_distillation(self, train_loader: DataLoader):
-        """Run progressive distillation through step counts"""
-        distillation_results = self.consistency_model.progressive_distillation(train_loader)
-        print(f"Progressive distillation completed: {distillation_results}")
-    
-    def save_checkpoint(self, path: str):
-        """Save training checkpoint with all models"""
-        checkpoint = {
-            'diffusion_model': self.diffusion_model.state_dict(),
-            'consistency_model': self.consistency_model.state_dict(),
-            'converter': self.converter.state_dict(),
-            'ema_model': self.ema_model.state_dict(),
-            'optimizer': self.optimizer.state_dict(),
-            'scheduler': self.scheduler.state_dict(),
-            'scaler': self.scaler.state_dict(),
-            'global_step': self.global_step,
-            'model_config': asdict(self.model_config),
-            'diffusion_config': asdict(self.diffusion_config),
-            'training_config': asdict(self.training_config),
-        }
-        torch.save(checkpoint, path)
-        print(f"Optimized checkpoint saved to {path}")
-    
-    def load_checkpoint(self, path: str):
-        """Load training checkpoint"""
-        checkpoint = torch.load(path, map_location=self.device)
-        self.diffusion_model.load_state_dict(checkpoint['diffusion_model'])
-        self.consistency_model.load_state_dict(checkpoint['consistency_model'])
-        self.converter.load_state_dict(checkpoint['converter'])
-        self.ema_model.load_state_dict(checkpoint['ema_model'])
-        self.optimizer.load_state_dict(checkpoint['optimizer'])
-        self.scheduler.load_state_dict(checkpoint['scheduler'])
-        if 'scaler' in checkpoint:
-            self.scaler.load_state_dict(checkpoint['scaler'])
-        self.global_step = checkpoint['global_step']
-        print(f"Optimized checkpoint loaded from {path}")
-
-# ============================================================================
-# INFERENCE & MARCHING CUBES WITH OPTIMIZATIONS
-# ============================================================================
-
-class OptimizedAircraftGenerator:
-    """Optimized inference engine with 4-step generation"""
-    
-    def __init__(self, checkpoint_path: str, device: torch.device = None):
-        self.device = device or torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-        
-        # Load checkpoint
-        checkpoint = torch.load(checkpoint_path, map_location=self.device)
-        
-        self.model_config = ModelConfig(**checkpoint['model_config'])
-        self.diffusion_config = DiffusionConfig(**checkpoint['diffusion_config'])
-        
-        self.diffusion_model = LatentDiffusionUNet(self.model_config, self.diffusion_config).to(self.device)
-        self.converter = LatentTo3DConverter(self.model_config.latent_dim, self.model_config.grid_resolution).to(self.device)
-        
-        # Load consistency model
-        self.consistency_model = ConsistencyModel(self.model_config, self.diffusion_config).to(self.device)
-        self.consistency_model.load_state_dict(checkpoint['consistency_model'])
-        
-        self.diffusion_model.load_state_dict(checkpoint['diffusion_model'])
-        self.converter.load_state_dict(checkpoint['converter'])
-        
-        self.noise_schedule = NoiseSchedule(self.diffusion_config).to(self.device)
-        
-        self.diffusion_model.eval()
-        self.converter.eval()
-        self.consistency_model.student_model.eval()
-    
-    @torch.no_grad()
-    def generate(self, design_spec: DesignSpec, num_steps: int = 4, guidance_scale: float = 7.5) -> torch.Tensor:
-        """
-        Generate aircraft design via 4-step consistency model.
-        250x faster than original 1000-step diffusion!
-        """
-        latent_shape = (1, self.model_config.latent_dim)
-        x_t = torch.randn(latent_shape, device=self.device)
-        
-        print(f"Generating with 4-step consistency model (vs 1000-step diffusion)")
-        
-        # Use fast 4-step consistency model
-        voxel_grid = self.consistency_model.fast_inference(latent_shape, num_steps=num_steps)
-        voxel_grid = torch.sigmoid(self.converter(voxel_grid))
-        print((voxel_grid.max().item(), voxel_grid.min().item()))
-        return voxel_grid.squeeze(0)
-
-    
-    def _postprocess_voxels(self, voxel_grid: torch.Tensor, min_component_size: int = 32) -> torch.Tensor:
-        """Light cleanup for exported voxel geometries."""
-        if voxel_grid.ndim == 4:
-            voxel_grid = voxel_grid.squeeze(0)
-        binary = (voxel_grid > 0.5).detach().cpu().numpy().astype(np.uint8)
-        try:
-            from scipy import ndimage
-            labels, n = ndimage.label(binary)
-            if n <= 1:
-                return torch.as_tensor(binary, dtype=voxel_grid.dtype, device=voxel_grid.device)
-            sizes = ndimage.sum(binary, labels, index=range(1, n + 1))
-            keep = {i + 1 for i, size in enumerate(sizes) if size >= min_component_size}
-            cleaned = np.isin(labels, list(keep)).astype(np.uint8)
-            if cleaned.sum() == 0:
-                cleaned = binary
-        except Exception:
-            cleaned = binary
-        return torch.as_tensor(cleaned, dtype=voxel_grid.dtype, device=voxel_grid.device)
-
-    def voxels_to_stl(self, voxel_grid: torch.Tensor, output_path: str, use_marching_cubes: bool = True):
-        """Convert voxel grid to STL file using marching cubes with optimizations"""
-        
-        # Convert to numpy
-        voxel_np = voxel_grid.cpu().numpy()
-        
-        # Threshold to get binary grid
-        binary_grid = (voxel_np > 0.5).astype(np.float32)
-        
-        if use_marching_cubes:
-            print("Applying marching cubes with adaptive mesh refinement...")
-            try:
-                # Dynamic level setting for stability
-                level = (voxel_np.min() + voxel_np.max()) / 2.0
-                
-                vertices, faces, normals, values = measure.marching_cubes(
-                    binary_grid,
-                    level=level,
-                    spacing=(1.0, 1.0, 1.0)
-                )
-
-                scale = float(self.config.lbm_config.physical_length_scale)
-                h = scale / float(self.config.base_grid_resolution)
-                vertices = vertices * h - (scale * 0.5) + (0.5 * h)
-                
-                print(f"Generated optimized mesh: {len(vertices)} vertices, {len(faces)} faces")
-                
-                # Simplify mesh if too complex for performance
-                if len(faces) > 10000:
-                    print(f"Simplifying mesh from {len(faces)} faces for performance")
-                    mesh = trimesh.Trimesh(vertices=vertices, faces=faces)
-                    try:
-                        # Use trimesh simplification
-                        simplified = mesh.simplify_quadratic_decimation(face_count=min(5000, len(mesh.faces)//2))
-                        vertices, faces = simplified.vertices, simplified.faces
-                        print(f"Simplified to: {len(vertices)} vertices, {len(faces)} faces")
-                    except Exception as e:
-                        print(f"Mesh simplification failed: {e}")
-                
-                self._write_stl(output_path, vertices, faces)
-                print(f"Optimized STL file written to {output_path}")
-            except Exception as e:
-                print(f"Marching cubes failed: {e}. Writing voxel representation instead.")
-                self._write_voxel_stl(output_path, binary_grid)
-        else:
-            self._write_voxel_stl(output_path, binary_grid)
-    
-    def _write_stl(self, path: str, vertices: np.ndarray, faces: np.ndarray):
-        """Write mesh to binary STL file with optimizations"""
-        with open(path, 'wb') as f:
-            # Header
-            f.write(b'\0' * 80)
-            # Number of triangles
-            f.write(np.uint32(len(faces)).tobytes())
-            
-            # Write each triangle
-            for face in faces:
-                v0, v1, v2 = vertices[face[0]], vertices[face[1]], vertices[face[2]]
-                normal = np.cross(v1 - v0, v2 - v0)
-                norm = np.linalg.norm(normal)
-                if norm > 1e-12:
-                    normal = normal / norm
-                else:
-                    normal = np.zeros(3)
-                
-                f.write(normal.astype(np.float32).tobytes())
-                f.write(v0.astype(np.float32).tobytes())
-                f.write(v1.astype(np.float32).tobytes())
-                f.write(v2.astype(np.float32).tobytes())
-                f.write(b'\0\0')  # Attribute byte count
-    
-    def _write_voxel_stl(self, path: str, binary_grid: np.ndarray):
-        """Write voxel grid as STL cubes with optimizations"""
-        triangles = []
-        
-        # Optimized voxel processing
-        for x in range(binary_grid.shape[0]):
-            for y in range(binary_grid.shape[1]):
-                for z in range(binary_grid.shape[2]):
-                    if binary_grid[x, y, z] > 0.5:
-                        # Create cube at this voxel
-                        vertices = np.array([
-                            [x, y, z], [x+1, y, z], [x+1, y+1, z], [x, y+1, z],
-                            [x, y, z+1], [x+1, y, z+1], [x+1, y+1, z+1], [x, y+1, z+1]
-                        ], dtype=np.float32)
-                        scale = float(self.config.lbm_config.physical_length_scale)
-                        h = scale / float(self.config.base_grid_resolution)
-                        vertices = vertices * h - (scale * 0.5) + (0.5 * h)
-                        
-                        # Cube face indices
-                        faces = [
-                            [0, 1, 2], [0, 2, 3], [4, 6, 5], [4, 7, 6],
-                            [0, 4, 5], [0, 5, 1], [2, 6, 7], [2, 7, 3],
-                            [0, 3, 7], [0, 7, 4], [1, 5, 6], [1, 6, 2]
-                        ]
-                        
-                        for face in faces:
-                            triangles.append(vertices[face])
-        
-        if triangles:
-            triangles = np.array(triangles)
-            vertices = triangles.reshape(-1, 3)
-            faces = np.arange(len(vertices)).reshape(-1, 3)
-            self._write_stl(path, vertices, faces)
-
-    def export_openfoam_case(self, voxel_grid: torch.Tensor, case_dir: str) -> Dict[str, Any]:
-        """Create a minimal OpenFOAM validation case around the exported geometry."""
-        case_path = Path(case_dir)
-        tri_surface = case_path / "constant" / "triSurface"
-        system = case_path / "system"
-        constant = case_path / "constant"
-        for p in (tri_surface, system, constant / "polyMesh", case_path / "0"):
-            p.mkdir(parents=True, exist_ok=True)
-
-        processed = self._postprocess_voxels(voxel_grid.unsqueeze(0)).squeeze(0) if voxel_grid.ndim == 3 else self._postprocess_voxels(voxel_grid)
-        stl_path = tri_surface / "design.stl"
-        self.voxels_to_stl(processed, str(stl_path), use_marching_cubes=True)
-
-        (system / "blockMeshDict").write_text("""FoamFile\n{\n    version 2.0;\n    format ascii;\n    class dictionary;\n    object blockMeshDict;\n}\nconvertToMeters 1;\nvertices\n(\n    (-5 -2 -2)\n    ( 5 -2 -2)\n    ( 5  2 -2)\n    (-5  2 -2)\n    (-5 -2  2)\n    ( 5 -2  2)\n    ( 5  2  2)\n    (-5  2  2)\n);\nblocks\n(\n    hex (0 1 2 3 4 5 6 7) (60 24 24) simpleGrading (1 1 1)\n);\nedges ( );\nboundary\n(\n    inlet { type patch; faces ((0 4 7 3)); }\n    outlet { type patch; faces ((1 2 6 5)); }\n    top { type patch; faces ((3 7 6 2)); }\n    bottom { type patch; faces ((0 1 5 4)); }\n    front { type symmetryPlane; faces ((0 3 2 1)); }\n    back { type symmetryPlane; faces ((4 5 6 7)); }\n);\nmergePatchPairs ( );\n""")
-        (system / "snappyHexMeshDict").write_text("""FoamFile
-{ version 2.0; format ascii; class dictionary; object snappyHexMeshDict; }
-castellatedMesh true;
-snap true;
-addLayers false;
-mergeTolerance 1e-6;
-geometry
-{ design.stl { type triSurfaceMesh; name design; } }
-castellatedMeshControls
-{
-    maxLocalCells 50000; maxGlobalCells 200000; minRefinementCells 0; nCellsBetweenLevels 2;
-    features ( ); refinementSurfaces { design { level (1 2); } }; refinementRegions { };
-    allowFreeStandingZoneFaces true; resolveFeatureAngle 30; locationInMesh (4 0 0);
-}
-snapControls { nSmoothPatch 3; tolerance 2.0; nSolveIter 30; nRelaxIter 5; }
-addLayersControls
-{
-    relativeSizes true;
-    layers { }
-    expansionRatio 1.0;
-    finalLayerThickness 0.3;
-    minThickness 0.1;
-    nGrow 0;
-    featureAngle 30;
-    nRelaxIter 3;
-    nSmoothSurfaceNormals 1;
-    nSmoothNormals 3;
-    nSmoothThickness 10;
-    maxFaceThicknessRatio 0.5;
-    maxThicknessToMedialRatio 0.3;
-    minMedialAxisAngle 90;
-    nBufferCellsNoExtrude 0;
-    nLayerIter 0;
-}
-meshQualityControls
-{
-    maxNonOrtho 65;
-    maxBoundarySkewness 20;
-    maxInternalSkewness 4;
-    maxConcave 80;
-    minVol 1e-13;
-    minTetQuality 1e-30;
-    minArea -1;
-    minTwist 0.02;
-    minDeterminant 0.001;
-    minFaceWeight 0.02;
-    minVolRatio 0.01;
-    minTriangleTwist -1;
-    nSmoothScale 4;
-    errorReduction 0.75;
-}
-""")
-        (system / "controlDict").write_text("""/*--------------------------------*- C++ -*----------------------------------*\\
-| =========                 |                                                 |
-| \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox           |
-|  \\    /   O peration     | Version:  v1912                                 |
-|   \\  /    A nd           | Website:  www.openfoam.com                      |
-|    \\/     M anipulation  |                                                 |
-\\*---------------------------------------------------------------------------*/
-FoamFile
-{
-    version     2.0;
-    format      ascii;
-    class       dictionary;
-    location    \"system\";
-    object      controlDict;
-}
-application     sonicFoam;
-startFrom       latestTime;
-startTime       0;
-stopAt          endTime;
-endTime         0.0027;
-deltaT          4e-08;
-writeControl    runTime;
-writeInterval   2e-04;
-purgeWrite      0;
-writeFormat     ascii;
-writePrecision  6;
-writeCompression off;
-timeFormat      general;
-timePrecision   6;
-runTimeModifiable true;
-""")
-        (system / "fvSchemes").write_text("""/*--------------------------------*- C++ -*----------------------------------*\\
-| =========                 |                                                 |
-| \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox           |
-|  \\    /   O operation     | Version:  v1912                                 |
-|   \\  /    A nd           | Website:  www.openfoam.com                      |
-|    \\/     M anipulation  |                                                 |
-\\*---------------------------------------------------------------------------*/
-FoamFile
-{
-    version     2.0;
-    format      ascii;
-    class       dictionary;
-    location    \"system\";
-    object      fvSchemes;
-}
-ddtSchemes
-{
-    default         Euler;
-}
-gradSchemes
-{
-    default         Gauss linear;
-    grad(U)         cellLimited Gauss linear 1;
-}
-divSchemes
-{
-    default         none;
-    div(phi,U)      Gauss limitedLinearV 1;
-    div(phi,e)      Gauss limitedLinear 1;
-    div(phid,p)     Gauss limitedLinear 1;
-    div(phiv,p)     Gauss limitedLinear 1;
-    div(phi,K)      Gauss limitedLinear 1;
-    div(phi,k)      Gauss upwind;
-    div(phi,epsilon) Gauss upwind;
-    div(((rho*nuEff)*dev2(T(grad(U))))) Gauss linear;
-}
-laplacianSchemes
-{
-    default         Gauss linear limited corrected 0.5;
-}
-interpolationSchemes
-{
-    default         linear;
-}
-snGradSchemes
-{
-    default         corrected;
-}
-""")
-        (system / "fvSolution").write_text("""/*--------------------------------*- C++ -*----------------------------------*\\
-| =========                 |                                                 |
-| \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox           |
-|  \\    /   O peration     | Version:  v1912                                 |
-|   \\  /    A nd           | Website:  www.openfoam.com                      |
-|    \\/     M anipulation  |                                                 |
-\\*---------------------------------------------------------------------------*/
-FoamFile
-{
-    version     2.0;
-    format      ascii;
-    class       dictionary;
-    location    \"system\";
-    object      fvSolution;
-}
-solvers
-{
-    \"rho.*\"
-    {
-        solver          diagonal;
-    }
-
-    \"p.*\"
-    {
-        solver          PBiCGStab;
-        preconditioner  DILU;
-        tolerance       1e-12;
-        relTol          0;
-    }
-
-    \"(U|e).*\"
-    {
-        $p;
-        tolerance       1e-9;
-    }
-
-    \"(k|epsilon).*\"
-    {
-        $p;
-        tolerance       1e-10;
-    }
-}
-PIMPLE
-{
-    nOuterCorrectors 1;
-    nCorrectors      2;
-    nNonOrthogonalCorrectors 0;
-}
-""")
-        (constant / "thermophysicalProperties").write_text("""/*--------------------------------*- C++ -*----------------------------------*\\
-| =========                 |                                                 |
-| \\      /  F ield         | OpenFOAM: The Open Source CFD Toolbox           |
-|  \\    /   O operation     | Version:  v1912                                 |
-|   \\  /    A nd           | Website:  www.openfoam.com                      |
-|    \\/     M anipulation  |                                                 |
-\\*---------------------------------------------------------------------------*/
-FoamFile
-{
-    version     2.0;
-    format      ascii;
-    class       dictionary;
-    location    \"constant\";
-    object      thermophysicalProperties;
-}
-thermoType
-{
-    type            hePsiThermo;
-    mixture         pureMixture;
-    transport       const;
-    thermo          hConst;
-    equationOfState perfectGas;
-    specie          specie;
-    energy          sensibleInternalEnergy;
-}
-mixture
-{
-    specie
-    {
-        molWeight       28.9;
-    }
-    thermodynamics
-    {
-        Cp              1005;
-        Hf              0;
-    }
-    transport
-    {
-        mu              0;
-        Pr              0.7;
-    }
-}
-""")
-        (constant / "turbulenceProperties").write_text("""FoamFile
-{ version 2.0; format ascii; class dictionary; object turbulenceProperties; }
-simulationType laminar;
-""")
-        (system / "forces").write_text("""FoamFile\n{ version 2.0; format ascii; class dictionary; object forces; }\ntype forces;\nfunctionObjectLibs (\"libforces.so\");\npatches (design);\nrho rho;\nrhoInf 1.225;\np p;\nU U;\nCofR (0 0 0);\n""")
-        (case_path / "0" / "U").write_text("""FoamFile\n{ version 2.0; format ascii; class volVectorField; object U; }\ndimensions [0 1 -1 0 0 0 0];\ninternalField uniform (80 0 0);\nboundaryField { inlet { type fixedValue; value uniform (80 0 0); } outlet { type pressureInletOutletVelocity; value uniform (80 0 0); } top { type slip; } bottom { type slip; } front { type symmetryPlane; } back { type symmetryPlane; } design { type noSlip; } }\n""")
-        (case_path / "0" / "p").write_text("""FoamFile\n{ version 2.0; format ascii; class volScalarField; object p; }\ndimensions [1 -1 -2 0 0 0 0];\ninternalField uniform 101325;\nboundaryField { inlet { type totalPressure; p0 uniform 101325; value uniform 101325; } outlet { type fixedValue; value uniform 101325; } top { type zeroGradient; } bottom { type zeroGradient; } front { type symmetryPlane; } back { type symmetryPlane; } design { type zeroGradient; } }\n""")
-        (case_path / "0" / "T").write_text("""FoamFile\n{ version 2.0; format ascii; class volScalarField; object T; }\ndimensions [0 0 0 1 0 0 0];\ninternalField uniform 300;\nboundaryField { inlet { type fixedValue; value uniform 300; } outlet { type zeroGradient; } top { type zeroGradient; } bottom { type zeroGradient; } front { type symmetryPlane; } back { type symmetryPlane; } design { type zeroGradient; } }\n""")
-        (case_path / "0" / "rho").write_text("""FoamFile\n{ version 2.0; format ascii; class volScalarField; object rho; }\ndimensions [1 -3 0 0 0 0 0];\ninternalField uniform 1.225;\nboundaryField { inlet { type fixedValue; value uniform 1.225; } outlet { type zeroGradient; } top { type zeroGradient; } bottom { type zeroGradient; } front { type symmetryPlane; } back { type symmetryPlane; } design { type zeroGradient; } }\n""")
-        (case_path / "OPENFOAM_EXPORT.md").write_text(
-            "OpenFOAM validation case generated from repo geometry export.\n"
-            f"OpenFOAM root: {OPENFOAM_ROOT}\n"
-            "Run blockMesh -> surfaceFeatureExtract -> snappyHexMesh -> sonicFoam with forces.\n"
-        )
-        return {"case_dir": str(case_path), "stl_path": str(stl_path), "openfoam_available": OPENFOAM_AVAILABLE}
-
-# ============================================================================
-# CLI INTERFACE
-# ============================================================================
-
 import click
+import torch
+import numpy as np
+import trimesh
+from pathlib import Path
+from torch.utils.data import DataLoader
+
+from config import (
+    ModelConfig, DiffusionConfig, TrainingConfig, CFDConfig, DesignSpec,
+    LBMPhysicsConfig, OPENFOAM_AVAILABLE, OPENFOAM_ROOT, OPENFOAM_BIN
+)
+from models import (
+    GroupedQueryAttention, GradientCheckpointingWrapper, SpatialAttention,
+    ResidualBlock3D, LatentDiffusionUNet, ConsistencyModel, LatentTo3DConverter,
+    NoiseSchedule
+)
+from data_utils import (
+    AircraftDesignDataset, ConnectivityLoss, GroundTruthExporter, AerodynamicLoss
+)
+from cfd_simulator import AdvancedCFDSimulator
+from trainer import OptimizedDiffusionTrainer
+from generator import OptimizedAircraftGenerator
+from mesh_utils import normalize_stl_mesh
+from utils import get_vram_limit_resolution, get_stl_adaptive_resolution
 
 @click.group()
 def cli():
     """Aircraft Structural Design via Diffusion Models + CFD (Fully Optimized)"""
-    print("🚀 TRM/HRM Recursive Style Implementation")
-    print("✨ Features: 4-step consistency, grouped-query attention, gradient checkpointing")
-    print("⚡ Performance: 60% VRAM savings, 50% KV-cache reduction, adaptive mesh refinement")
-    print("🎯 CFD: FluidX3D + GPU LBM solver with SoA layout and 256-thread blocks")
-    print("🔄 Pipeline: CFD computation overlapped with diffusion sampling")
+    print("🚀 TRM/HRM Recursive Style Implementation (Modular)")
     pass
 
 @cli.command()
-@click.option('--num-epochs', default=100, help='Number of training epochs')
-@click.option('--batch-size', default=4, help='Batch size')
-@click.option('--learning-rate', default=2e-4, help='Learning rate')
-@click.option('--latent-dim', default=16, help='Latent dimension')
-@click.option('--precision', default='float32', help='Precision: float64, float32, float16, bfloat16, float8')
-@click.option('--disconnection-penalty', default=30.0, help='Penalty for disconnected voxels')
-@click.option('--num-samples', default=500, help='Number of training samples')
-@click.option('--resume-from', default=None, help='Resume from checkpoint')
-@click.option('--save-dir', default='./checkpoints', help='Directory to save checkpoints')
-@click.option('--enable-consistency', is_flag=True, default=True, help='Enable 4-step consistency model')
-@click.option('--enable-pipeline', is_flag=True, default=True, help='Enable pipeline parallelism')
-@click.option('--enable-checkpointing', is_flag=True, default=True, help='Enable gradient checkpointing')
-@click.option('--enable-compile', is_flag=True, default=False, help='Enable torch.compile optimization')
-@click.option('--solver', default='D3Q27', help='CFD solver type: D3Q27')
-def get_vram_limit_resolution(max_usage=0.9, target_res=1024):
-    """Estimate max LBM resolution based on available VRAM"""
-    if not torch.cuda.is_available():
-        return 64
-
-    try:
-        total_vram = torch.cuda.get_device_properties(0).total_memory
-        usable_vram = total_vram * max_usage
-
-        # D3Q27 memory: ~250 bytes per cell (populations + overhead)
-        bytes_per_cell = 250
-        max_cells = usable_vram / bytes_per_cell
-        max_res = int(max_cells ** (1/3))
-
-        return min(max(max_res, 32), target_res)
-    except Exception:
-        return 128
-
-def get_stl_adaptive_resolution(stl_path):
-    """Determine optimal resolution based on STL complexity"""
-    try:
-        mesh = trimesh.load(stl_path)
-        # complexity-based heuristic
-        complexity_res = int((len(mesh.faces) / 2) ** (1/3) * 4)
-        vram_res = get_vram_limit_resolution()
-        return min(max(complexity_res, 64), vram_res)
-    except Exception:
-        return get_vram_limit_resolution()
-
+@click.option('--num-epochs', default=100)
+@click.option('--batch-size', default=4)
+@click.option('--learning-rate', default=2e-4)
+@click.option('--latent-dim', default=16)
+@click.option('--precision', default='float32')
+@click.option('--disconnection-penalty', default=30.0)
+@click.option('--num-samples', default=500)
+@click.option('--resume-from', default=None)
+@click.option('--save-dir', default='./checkpoints')
+@click.option('--enable-consistency', is_flag=True, default=True)
+@click.option('--enable-pipeline', is_flag=True, default=True)
+@click.option('--enable-checkpointing', is_flag=True, default=True)
+@click.option('--enable-compile', is_flag=True, default=False)
+@click.option('--solver', default='D3Q27')
 def train(num_epochs, batch_size, learning_rate, latent_dim, precision, disconnection_penalty, 
           num_samples, resume_from, save_dir, enable_consistency, enable_pipeline, 
           enable_checkpointing, enable_compile, solver):
-    """Train the optimized diffusion model with all TRM/HRM features"""
-    import os
-    import logging
-
-    # Set environment variables BEFORE importing torch
-    os.environ["TORCHDYNAMO_VERBOSE"] = "1"
-    os.environ["TORCH_LOGS"] = "+dynamo,+inductor,output_code,graph_code,graph_breaks,guards"
-
-    # Now import torch
-    import torch
-
-    # Also set the logging API for maximum verbosity
-    torch._logging.set_logs(
-        dynamo=logging.DEBUG,
-        aot=logging.DEBUG,
-        inductor=logging.DEBUG,
-        output_code=True,
-        graph_code=True,
-        graph_breaks=True,
-        guards=True,
-        recompiles=True
-    )
-
-
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-    print(f"Using device: {device}")
-    
-    if torch.cuda.is_available():
-        print(f"GPU Memory: {torch.cuda.get_device_properties(0).total_memory / 1e9:.2f} GB")
-        print(f"Memory Optimization: 60% VRAM savings enabled")
-    
-    # Load checkpoint if resuming
-    model_config_override = None
-    if resume_from:
-        checkpoint = torch.load(resume_from, map_location=device)
-        model_config_override = ModelConfig(**checkpoint['model_config'])
-        print(f"Loaded model config from checkpoint: latent_dim={model_config_override.latent_dim}")
-
-    # Create directories
     Path(save_dir).mkdir(parents=True, exist_ok=True)
 
-    # Optimized configs
-    model_config = model_config_override if model_config_override else ModelConfig(
-        latent_dim=latent_dim,
-        attention_groups=4,  # Grouped-query attention
-        enable_gradient_checkpointing=enable_checkpointing,
-        use_torch_compile=enable_compile  # Respect the enable-compile flag
-    )
-    
-    diffusion_config = DiffusionConfig(
-        teacher_steps=1000,
-        student_steps=4  # 4-step consistency model
-    )
-    
-    training_config = TrainingConfig(
-        num_epochs=num_epochs,
-        batch_size=batch_size,
-        learning_rate=learning_rate,
-        disconnection_penalty=disconnection_penalty,
-        precision=precision,
-        enable_pipeline_parallelism=enable_pipeline
-    )
-    
-    # VRAM limit for training (can start small and grow)
-    base_resolution = 32
+    model_config = ModelConfig(latent_dim=latent_dim, attention_groups=4, enable_gradient_checkpointing=enable_checkpointing, use_torch_compile=enable_compile)
+    diffusion_config = DiffusionConfig(teacher_steps=1000, student_steps=4)
+    training_config = TrainingConfig(num_epochs=num_epochs, batch_size=batch_size, learning_rate=learning_rate, disconnection_penalty=disconnection_penalty, precision=precision, enable_pipeline_parallelism=enable_pipeline)
+    cfd_config = CFDConfig(base_grid_resolution=16 if solver == "D3Q27" else 32, solver_type=solver)
 
-    cfd_config = CFDConfig(
-        base_grid_resolution=base_resolution,
-        adaptive_cells_target=5000,
-        solver_type=solver
-    )
-
-    # Dataset - start at 32 and grow to target
-    dataset = AircraftDesignDataset(
-        num_samples=num_samples,
-        grid_size=32,
-        latent_dim=model_config.latent_dim,
-        target_grid_size=1024
-    )
+    dataset = AircraftDesignDataset(num_samples=num_samples, grid_size=32, latent_dim=model_config.latent_dim, target_grid_size=1024)
     train_loader = DataLoader(dataset, batch_size=batch_size, shuffle=True, num_workers=2)
 
-    # Optimized trainer
-    trainer = OptimizedDiffusionTrainer(
-        model_config, diffusion_config, training_config, cfd_config, device=device
-    )
-
+    trainer = OptimizedDiffusionTrainer(model_config, diffusion_config, training_config, cfd_config, device=device)
     if resume_from:
         trainer.load_checkpoint(resume_from)
-        print(f"Resumed from {resume_from}")
-    
-    print("\n" + "="*60)
-    print("🚀 STARTING OPTIMIZED TRAINING")
-    print("="*60)
-    print(f"✨ 4-step consistency model: {enable_consistency}")
-    print(f"🔄 Pipeline parallelism: {enable_pipeline}")
-    print(f"💾 Gradient checkpointing: {enable_checkpointing}")
-    print(f"⚡ torch.compile optimization: {enable_compile}")
-    print(f"🎯 Adaptive mesh refinement: ~5k cells (vs 32³ = 32k)")
-    print(f"🧠 Grouped-query attention: 50% KV-cache reduction")
-    print("="*60)
-    
-    # Train with optimizations
     trainer.train(train_loader)
-    
-    # Save final model
-    final_checkpoint = os.path.join(save_dir, 'final_optimized_model.pt')
-    trainer.save_checkpoint(final_checkpoint)
-    print(f"\n🎉 Training complete! Final optimized model saved to {final_checkpoint}")
-    print(f"📊 Achieved: 250x speedup (4-step vs 1000-step), 60% VRAM savings, 50% memory reduction")
+    trainer.save_checkpoint(os.path.join(save_dir, 'final_optimized_model.pt'))
 
 @cli.command()
-@click.option('--design', required=True, help='Path to .npy design file or .stl file')
-@click.option('--cfd-steps', default=500, help='Number of CFD simulation steps')
-@click.option('--solver', default='D3Q27', help='CFD solver type')
+@click.option('--design', required=True)
+@click.option('--cfd-steps', default=500)
+@click.option('--solver', default='D3Q27')
 def evaluate(design, cfd_steps, solver):
-    """Evaluate design with high-fidelity CFD and mesh normalization"""
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-
     if design.endswith('.stl'):
         res = get_stl_adaptive_resolution(design)
-        print(f"🚀 Evaluating STL: {design} at resolution {res}^3")
-
         cfd_config = CFDConfig(base_grid_resolution=res, solver_type=solver)
         simulator = AdvancedCFDSimulator(cfd_config, device)
-
-        # Load and normalize STL to fit simulation domain
         mesh = trimesh.load(design)
-
-        # Center and scale to fit [0, 1]^3 with 10% padding
-        extents = mesh.extents
-        scale = 0.8 / max(extents)
-        mesh.apply_translation(-mesh.centroid)
-        mesh.apply_scale(scale)
-        mesh.apply_translation([0.5, 0.5, 0.5])
-
-        # Voxelize
-        voxel_grid = mesh.voxelized(1.0 / res).matrix
-
-        # Sync with simulator grid
+        mesh = normalize_stl_mesh(mesh)
+        voxel_grid_np = mesh.voxelized(1.0 / res).matrix
         geometry = torch.zeros((res, res, res), device=device)
-        d, h, w = min(res, voxel_grid.shape[0]), min(res, voxel_grid.shape[1]), min(res, voxel_grid.shape[2])
-        geometry[:d, :h, :w] = torch.from_numpy(voxel_grid[:d, :h, :w].astype(np.float32)).to(device)
+        d, h, w = min(res, voxel_grid_np.shape[0]), min(res, voxel_grid_np.shape[1]), min(res, voxel_grid_np.shape[2])
+        geometry[:d, :h, :w] = torch.from_numpy(voxel_grid_np[:d, :h, :w].astype(np.float32)).to(device)
     else:
         voxel_grid_np = np.load(design)
         res = voxel_grid_np.shape[0]
-        print(f"🚀 Evaluating design: {design} at resolution {res}^3")
         geometry = torch.from_numpy(voxel_grid_np).to(device)
         cfd_config = CFDConfig(base_grid_resolution=res, solver_type=solver)
         simulator = AdvancedCFDSimulator(cfd_config, device)
-
     results = simulator.simulate_aerodynamics(geometry, steps=cfd_steps)
-    print(f"\n📊 EVALUATION RESULTS for {design}:")
     for k, v in results.items():
-        print(f"  • {k}: {v:.6f}")
+        if isinstance(v, (int, float)): print(f"  • {k}: {v:.6f}")
 
 @cli.command()
-@click.option('--checkpoint', required=True, help='Path to model checkpoint')
-@click.option('--output', default='aircraft_optimized.stl', help='Output STL file path')
-@click.option('--target-speed', default=7.0, help='Target aircraft speed (m/s)')
-@click.option('--num-steps', default=4, help='Number of diffusion steps for generation (4 for consistency)')
-@click.option('--use-marching-cubes', is_flag=True, default=True, help='Use marching cubes for STL conversion')
-@click.option('--solver', default='D3Q27', help='CFD solver type: D3Q27')
-def generate(checkpoint, output, target_speed, num_steps, use_marching_cubes, solver):
-    """Generate aircraft design using optimized 4-step consistency model"""
-    
+@click.option('--stl', required=True)
+@click.option('--steps', default=500)
+def accuracy_benchmark(stl, steps):
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-    print(f"Using device: {device}")
-    
-    if not os.path.exists(checkpoint):
-        print(f"Error: Checkpoint not found at {checkpoint}")
-        sys.exit(1)
-    
-    # Load optimized generator
-    print(f"Loading optimized checkpoint from {checkpoint}...")
-    generator = OptimizedAircraftGenerator(checkpoint, device=device)
-    
-    # Design specification
-    design_spec = DesignSpec(
-        target_speed=target_speed,
-        space_weight=0.33,
-        drag_weight=0.33,
-        lift_weight=0.34
-    )
-    
-    # Generate with 4-step consistency model
-    print(f"🚀 Generating aircraft design with 4-step consistency model...")
-    print(f"⚡ Speedup: 250x faster than 1000-step diffusion!")
-    voxel_grid = generator.generate(design_spec, num_steps=num_steps)
-    
-    print(f"Generated voxel grid shape: {voxel_grid.shape}")
-    print(f"Occupied voxels: {(voxel_grid > 0.5).sum().item()} / {np.prod(voxel_grid.shape)}")
-    
-    # Export to optimized STL
-    print(f"Converting to optimized STL mesh with adaptive refinement...")
-    generator.voxels_to_stl(voxel_grid, output, use_marching_cubes=use_marching_cubes)
+    print(f"\n🚀 CANONICAL ACCURACY BENCHMARK: {stl}")
+    for corr in [True, False]:
+        print(f"\n📊 Running {'with' if corr else 'WITHOUT'} Shape-Drag Correction...")
+        conf = CFDConfig(base_grid_resolution=64)
+        conf.lbm_config.use_shape_drag_correction = corr
+        sim = AdvancedCFDSimulator(conf, device)
+        mesh = trimesh.load(stl)
+        mesh = normalize_stl_mesh(mesh)
+        vg = mesh.voxelized(1.0/64).matrix
+        geom = torch.zeros((64, 64, 64), device=device)
+        d, h, w = min(64, vg.shape[0]), min(64, vg.shape[1]), min(64, vg.shape[2])
+        geom[:d, :h, :w] = torch.from_numpy(vg[:d, :h, :w].astype(np.float32)).to(device)
+        res = sim.simulate_aerodynamics(geom, steps=steps)
+        print(f"  Drag Coeff: {res['drag_coefficient']:.6f}, PINN Ready: {res.get('pinn_ready')}")
 
-    # Add final CFD analysis
-    print(f"🚀 Running final CFD analysis with {solver} solver...")
+@cli.command()
+@click.option('--checkpoint', required=True)
+@click.option('--output', default='aircraft_optimized.stl')
+@click.option('--target-speed', default=7.0)
+@click.option('--num-steps', default=4)
+@click.option('--use-marching-cubes', is_flag=True, default=True)
+@click.option('--solver', default='D3Q27')
+def generate(checkpoint, output, target_speed, num_steps, use_marching_cubes, solver):
+    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+    generator = OptimizedAircraftGenerator(checkpoint, device=device)
+    design_spec = DesignSpec(target_speed=target_speed)
+    voxel_grid = generator.generate(design_spec, num_steps=num_steps)
+    generator.save_stl(voxel_grid, output, use_marching_cubes=use_marching_cubes)
     cfd_config = CFDConfig(solver_type=solver)
     simulator = AdvancedCFDSimulator(cfd_config, device)
-    results = simulator.simulate_aerodynamics(voxel_grid, steps=1000)
-    print("CFD Analysis Results:")
-    print(f"  Drag Coefficient: {results['drag_coefficient']}")
-    print(f"  Lift Coefficient: {results['lift_coefficient']}")
+    results = simulator.simulate_aerodynamics(voxel_grid, steps=100)
+    print(f"Drag: {results['drag_coefficient']}, Lift: {results['lift_coefficient']}")
 
 @cli.command()
-@click.option('--checkpoint', required=True, help='Path to model checkpoint')
-@click.option('--output-dir', default='./generations_optimized', help='Output directory for generated designs')
-@click.option('--num-designs', default=5, help='Number of designs to generate')
+@click.option('--checkpoint', required=True)
+@click.option('--output-dir', default='./generations_optimized')
+@click.option('--num-designs', default=5)
 def batch_generate(checkpoint, output_dir, num_designs):
-    """Generate multiple aircraft designs using optimized pipeline"""
-    
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
     Path(output_dir).mkdir(parents=True, exist_ok=True)
-    
-    print(f"Using device: {device}")
-    print(f"Loading optimized checkpoint from {checkpoint}...")
     generator = OptimizedAircraftGenerator(checkpoint, device=device)
-    
-    design_spec = DesignSpec(target_speed=50.0)
-    
-    print(f"\n🚀 Generating {num_designs} optimized aircraft designs...")
-    print(f"⚡ Using 4-step consistency model with pipeline parallelism")
-    
     for i in range(num_designs):
-        print(f"\n🎨 Generating optimized design {i+1}/{num_designs}...")
-        voxel_grid = generator.generate(design_spec, num_steps=4)
-        
-        output_path = os.path.join(output_dir, f'aircraft_optimized_{i+1:03d}.stl')
-        generator.voxels_to_stl(voxel_grid, output_path, use_marching_cubes=True)
-
-@cli.command()
-def performance_benchmark():
-    """Benchmark all optimizations"""
-    print("\n🚀 TRM/HRM RECURSIVE STYLE PERFORMANCE BENCHMARK")
-    print("="*60)
-    
-    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-    print(f"Device: {device}")
-    
-    if torch.cuda.is_available():
-        props = torch.cuda.get_device_properties(0)
-        print(f"GPU: {props.name}")
-        print(f"Total Memory: {props.total_memory / 1e9:.2f} GB")
-        print(f"Compute Capability: {props.major}.{props.minor}")
-    
-    print("\n✨ OPTIMIZATION FEATURES:")
-    print("• 4-step consistency model: 250x speedup vs 1000-step diffusion")
-    print("• Grouped-query attention: 50% KV-cache reduction")
-    print("• Gradient checkpointing: 60% VRAM savings")
-    print("• torch.compile: Kernel fusion optimization")
-    print("• Adaptive mesh refinement: ~5k cells vs 32³ (85% reduction)")
-    print("• GPU LBM solver: SoA layout with 256-thread blocks")
-    print("• Pipeline parallelism: CFD + diffusion overlap")
-    
-    print("\n🎯 EXPECTED PERFORMANCE GAINS:")
-    print("• Inference Speed: 250x faster (4 steps vs 1000 steps)")
-    print("• Memory Usage: 60% VRAM reduction")
-    print("• Attention Memory: 50% KV-cache reduction")
-    print("• CFD Grid Size: 85% fewer cells")
-    print("• Training Throughput: 2-3x improvement")
-    
-    print("\n📊 BENCHMARK COMPLETE")
-    print("All TRM/HRM optimizations successfully implemented! 🎉")
+        voxel_grid = generator.generate(DesignSpec(target_speed=50.0))
+        generator.save_stl(voxel_grid, os.path.join(output_dir, f'aircraft_{i+1:03d}.stl'))
 
 @cli.command()
 def info():
-    """Print system and optimization information"""
-    print(f"\n🚀 TRM/HRM Recursive Style Implementation")
-    print(f"PyTorch version: {torch.__version__}")
-    print(f"CUDA available: {torch.cuda.is_available()}")
-    
-    if torch.cuda.is_available():
-        print(f"CUDA device: {torch.cuda.get_device_name(0)}")
-        print(f"CUDA capability: {torch.cuda.get_device_capability(0)}")
-        print(f"Total GPU memory: {torch.cuda.get_device_properties(0).total_memory / 1e9:.2f} GB")
-        print(f"Allocated GPU memory: {torch.cuda.memory_allocated(0) / 1e9:.2f} GB")
-        print(f"Reserved GPU memory: {torch.cuda.memory_reserved(0) / 1e9:.2f} GB")
-    
-    print(f"\n✨ OPTIMIZATION STATUS:")
-    print(f"• 4-step consistency model: ✅ ENABLED")
-    print(f"• Grouped-query attention: ✅ ENABLED (4 groups)")
-    print(f"• Gradient checkpointing: ✅ ENABLED")
-    print(f"• torch.compile optimization: ✅ ENABLED")
-    print(f"• Adaptive mesh refinement: ✅ ENABLED (~5k cells)")
-    print(f"• GPU LBM solver: ✅ ENABLED (SoA layout, 256-thread blocks)")
-    print(f"• Pipeline parallelism: ✅ ENABLED")
-    print(f"• FluidX3D integration: ✅ ENABLED")
+    print(f"PyTorch: {torch.__version__}, CUDA: {torch.cuda.is_available()}")
 
 if __name__ == '__main__':
     cli()
